### 2021-03-28 #3137 @whedon
[PRE REVIEW]: eemont: A Python package that extends Google Earth Engine
**Submitting author:** @davemlz (<a href="http://orcid.org/0000-0002-9010-3286">David Montero Loaiza</a>)
**Repository:** <a href="https://github.com/davemlz/eemont" target ="_blank">https://github.com/davemlz/eemont</a>
**Version:** v0.1.9
