### 2021-04-04 f1fa8b4
VIIRS added to scale
