### 2021-04-04 474df45
VIIRS added to maskClouds
