+++ eemont/image.py
@@ -774,7 +774,7 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
         if maskShadows:
             cloud = cloud.Or(qa.bitwiseAnd(1 << 3))
         mask2 = args.mask().reduce(ee.Reducer.min());
-        return args.updateMask(cloud.Not()).updateMask(mask2);
+        return args.updateMask(cloud.Not()).updateMask(mask2)
     
     def MOD09GA(args):
         qa = args.select('state_1km')
@@ -783,7 +783,7 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 2).eq(0))
         if maskCirrus:
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 8).eq(0))        
-        return args.updateMask(notCloud);
+        return args.updateMask(notCloud)
     
     def MCD15A3H(args):
         qa = args.select('FparExtra_QC')
@@ -792,7 +792,7 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 6).eq(0))
         if maskCirrus:
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 4).eq(0))        
-        return args.updateMask(notCloud);
+        return args.updateMask(notCloud)
     
     def MOD09Q1(args):
         qa = args.select('State')
@@ -801,7 +801,7 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 2).eq(0))
         if maskCirrus:
             notCloud = notCloud.And(qa.bitwiseAnd(1 << 8).eq(0))        
-        return args.updateMask(notCloud);
+        return args.updateMask(notCloud)
         
     def MOD09A1(args):
         qa = args.select('StateQA')
@@ -832,6 +832,24 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
         notCloud = qa.eq(0)
         return args.updateMask(notCloud)
     
+    def VNP09GA(args):
+        qf1 = args.select('QF1')
+        qf2 = args.select('QF2')
+        notCloud = qf1.bitwiseAnd(1 << 2).eq(0)
+        if maskShadows:
+            notCloud = notCloud.And(qf2.bitwiseAnd(1 << 3).eq(0))
+        if maskCirrus:
+            notCloud = notCloud.And(qf2.bitwiseAnd(1 << 6).eq(0))
+            notCloud = notCloud.And(qf2.bitwiseAnd(1 << 7).eq(0))
+        return args.updateMask(notCloud)
+    
+    def VNP13A1(args):
+        qa = args.select('pixel_reliability')
+        notCloud = qa.neq(9)
+        if maskShadows:
+            notCloud = notCloud.And(qa.neq(7))
+        return args.updateMask(notCloud)
+    
     lookup = {
         'COPERNICUS/S3': S3,
         'COPERNICUS/S2': S2,
@@ -855,7 +873,9 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
         'MODIS/006/MYD16A2': MOD16A2,
         'MODIS/006/MYD13Q1': MOD13Q1A1,
         'MODIS/006/MYD13A1': MOD13Q1A1,
-        'MODIS/006/MYD13A2': MOD13A2
[…]
