+++ docs/changelog.rst
@@ -1,6 +1,21 @@
 Changelog
 ============
 
+v0.1.10
+--------------
+
+Improvements
+~~~~~~~~~~~~~~~~~~~~~~
+
+- The :code:`maskClouds()` extended method for ee.Image and ee.ImageCollection classes now supports the following platforms:
+   
+   - `VNP09GA: VIIRS Surface Reflectance Daily 500m and 1km <https://developers.google.com/earth-engine/datasets/catalog/NOAA_VIIRS_001_VNP09GA?hl=en>`_
+   - `VNP13A1: VIIRS Vegetation Indices 16-Day 500m <https://developers.google.com/earth-engine/datasets/catalog/NOAA_VIIRS_001_VNP13A1?hl=en>`_
+- The :code:`scale()` extended method for ee.Image and ee.ImageCollection classes now supports the following platforms:
+   
+   - `VNP09GA: VIIRS Surface Reflectance Daily 500m and 1km <https://developers.google.com/earth-engine/datasets/catalog/NOAA_VIIRS_001_VNP09GA?hl=en>`_
+   - `VNP13A1: VIIRS Vegetation Indices 16-Day 500m <https://developers.google.com/earth-engine/datasets/catalog/NOAA_VIIRS_001_VNP13A1?hl=en>`_
+
 v0.1.9
 --------------
 
