### 2021-04-05 #3137 @kbarnhart
@davemlz I'll now start actively handling this submission, beginning by looking for reviewers. I have some thoughts on suitable reviewers but would also be interested in people you think would be appropriate for this submission.

Please indicate here people you think are appropriate (without tagging them with an @). [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). Do not feel limited to this list.

In addition, based on looking through the package, it looks like you've created a nice set of tutorials, but instructions for using them are not provided. I'd recommend adding some text to the readme/documentation that points a user to the tutorials.
