+++ docs/conf.py
@@ -41,6 +41,9 @@ extensions = [
     'sphinx.ext.autosectionlabel'
 ]
 
+autodoc_default_flags = ['members']
+autosummary_generate = True
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
@@ -55,63 +58,33 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+#html_theme = 'sphinx_rtd_theme'
 #html_theme = "insegel"
 #html_theme = "divio_docs_theme"
 #html_theme = "karma_sphinx_theme"
 #html_theme = "sphinx_materialdesign_theme"
+html_theme = "pydata_sphinx_theme"
 
-#html_theme_options = {
-    # Specify a list of menu in Header.
-    # Tuples forms:
-    #  ('Name', 'external url or path of pages in the document', boolean, 'icon name')
-    #
-    # Third argument:
-    # True indicates an external link.
-    # False indicates path of pages in the document.
-    #
-    # Fourth argument:
-    # Specify the icon name.
-    # For details see link.
-    # https://material.io/icons/
-#    'header_links' : [
-#        ('Home', 'index', False, 'home'),
-#        ("GitHub", "https://github.com/davemlz/eemont", True, 'link')
-#    ],
-
-    # Customize css colors.
-    # For details see link.
-    # https://getmdl.io/customize/index.html
-    #
-    # Values: amber, blue, brown, cyan deep_orange, deep_purple, green, grey, indigo, light_blue,
-    #         light_green, lime, orange, pink, purple, red, teal, yellow(Default: indigo)
-#    'primary_color': 'blue',
-    # Values: Same as primary_color. (Default: pink)
-#    'accent_color': 'pink',
-
-    # Customize layout.
-    # For details see link.
-    # https://getmdl.io/components/index.html#layout-section
-#    'fixed_drawer': False,
-#    'fixed_header': True,
-#    'header_waterfall': False,
-#    'header_scroll': False,
-
-    # Render title in header.
-    # Values: True, False (Default: False)
-#    'show_header_title': False,
-    # Render title in drawer.
-    # Values: True, False (Default: True)
-#    'show_drawer_title': True,
-    # Render footer.
-    # Values: True, False (Default: True)
-#    'show_footer': True
-#}
+html_theme_options = {    
+    "icon_links": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/davemlz/eemont",
+            "icon": "fab fa-github-square",
+        },
+        {
+            "name": "Twitter",
+            "url": "https://twitter.com/dmlmont",
+            "icon": "fab fa-twitter-square",
+        },
+    ],    
+}
 
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+html_logo = "_static/logo.png"
 
-html_sidebars = { '**': ['globaltoc.html', 'relations.html', 'sourcelink.html', 'searchbox.html'] }
+#html_sidebars = { '**': ['globaltoc.html', 'relations.html', 'sourcelink.html', 'searchbox.html'] }
