+++ docs/classes/stubs/eemont.imagecollection.maskClouds.rst
@@ -1,0 +1,6 @@
+eemont.imagecollection.maskClouds
+=================================
+
+.. currentmodule:: eemont.imagecollection
+
+.. autofunction:: maskClouds
