+++ docs/classes/stubs/eemont.geometry.MultiPointFromQuery.rst
@@ -1,0 +1,6 @@
+eemont.geometry.MultiPointFromQuery
+===================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: MultiPointFromQuery
