+++ docs/conf.py
@@ -58,12 +58,12 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-#html_theme = 'sphinx_rtd_theme'
+html_theme = 'sphinx_rtd_theme'
 #html_theme = "insegel"
 #html_theme = "divio_docs_theme"
 #html_theme = "karma_sphinx_theme"
 #html_theme = "sphinx_materialdesign_theme"
-html_theme = "pydata_sphinx_theme"
+#html_theme = "pydata_sphinx_theme"
 
 html_theme_options = {    
     "icon_links": [
