+++ docs/docs-requirements.txt
@@ -1,3 +1,4 @@
 recommonmark
 sphinxcontrib-autoprogram
 pydata-sphinx-theme
+google-auth<2.0dev
