+++ docs/docs-requirements.txt
@@ -2,4 +2,3 @@ sphinx==3.1.2
 recommonmark
 sphinxcontrib-autoprogram
 pydata-sphinx-theme
-google-auth
