+++ .readthedocs.yaml
@@ -3,7 +3,7 @@
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
 # Required
-version: 2
+# version: 2
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -11,6 +11,5 @@ sphinx:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.7
    install:
    - requirements: docs/docs-requirements.txt
