+++ docs/docs-requirements.txt
@@ -1,4 +1,4 @@
-sphinx>=1.4, !=1.5.4
+# sphinx>=1.4, !=1.5.4
 recommonmark
 sphinxcontrib-autoprogram
 pydata-sphinx-theme
