+++ .readthedocs.yaml
@@ -5,10 +5,6 @@
 # Required
 version: 2
 
-# Build documentation in the docs/ directory with Sphinx
-sphinx:
-   configuration: docs/conf.py
-
 # Optionally set the version of Python and requirements required to build your docs
 python:
    version: 3
