+++ docs/docs-requirements.txt
@@ -1,4 +1,5 @@
 sphinx
+pydata-sphinx-theme
 numpydoc
 recommonmark
 pandas
