+++ docs/classes/eeimagecollection.rst
@@ -3,13 +3,7 @@ ee.ImageCollection
 
 Extended methods for the ee.ImageCollection class:
 
-.. currentmodule:: eemont.imagecollection
-.. autosummary::
-   :toctree: stubs
-
-   closest 
-   getTimeSeriesByRegion
-   getTimeSeriesByRegions
-   index
-   maskClouds
-   scale
+.. automodule:: eemont.imagecollection
+   :members:
+   :undoc-members:
+   :show-inheritance:
