+++ docs/classes/eegeometry.rst
@@ -3,11 +3,7 @@ ee.Geometry
 
 Extended methods for the ee.Geometry class:
 
-.. currentmodule:: eemont.geometry
-
-.. autosummary::
-   :toctree: stubs
-
-   BBoxFromQuery
-   PointFromQuery
-   MultiPointFromQuery
+.. automodule:: eemont.geometry
+   :members:
+   :undoc-members:
+   :show-inheritance:
