### 2021-04-08 #3137 @elbeejay
@kbarnhart I'm happy to review this with the caveat that I don't know how to use R (yet), so I won't be able to test the `rgee` R-compatibility of this package.
