### 2021-04-12 #3137 @kbarnhart
@giswqs, @elbeejay, and @patrickcgray thanks for your enthusiasm to review this.

@giswqs and/or @patrickcgray would either of you be able to review the `rgee` portion of the package ([described here](https://eemont.readthedocs.io/en/0.1.9/guide/eemontR.html)).

Once I confirm that someone is able to review the R portion I'll add you all as reviewers and start the main review issue.
