### 2021-04-12 #3137 @giswqs
@kbarnhart Yes, I can review the rgee portion of the package.
