### 2021-04-12 #3168 @kbarnhart
@giswqs, @elbeejay, @patrickcgray - thanks for agreeing to review this submission to JOSS. In the prior comments there is a checklist that you can use to guide you through your review.

We have an automatic reminder set up in two weeks to ask you how the review is going. At present we request that reviewers complete their reviews within 6 weeks. JOSS is trying to be mindful of changes people have experienced due to COVID-19.

As you work through your review, if there are any issues that come up, please make an issue in the [eemont](https://github.com/davemlz/eemont) repository, and link to this issue (openjournals/joss-reviews/issues/3168). That way most of the discussion can occur on in-repo issues.

If you have any questions, please let me know (tag me here or email krbarnhart@usgs.gov).
