### 2021-04-13 davemlz/eemont#4 @elbeejay
[JOSS Review] Contributor instructions
Hi, could you please add some instructions for contributors on how they might go about adding to this package? For example is there a particular code style, doc-string style that this package adheres to? Would you prefer issues opened before pull requests?

Thanks!

This comment relates to the ongoing [JOSS review](https://github.com/openjournals/joss-reviews/issues/3168) of this package, comment pertains to the "Documentation" section of the JOSS reviewer guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
