### 2021-04-13 davemlz/eemont#5 @elbeejay
[JOSS Review] Package test automation/instructions
Looks like there is a set of [tests](https://github.com/davemlz/eemont/tree/master/tests) that have been written for this package. Can you either automate these via some sort of continuous integration, or provide detailed instructions in the documentation so that an individual can run the tests locally?

I would consider adding a "developer installation" set of instructions to the documentation so that an aspiring contributor could follow steps to clone, install, locally test and make eventually make contributions to the project, but I leave that decision up to you.

Thanks.

This comment relates to the ongoing [JOSS review](https://github.com/openjournals/joss-reviews/issues/3168) of this package, comment pertains to the "Documentation" section of the JOSS reviewer guidelines:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
