### 2021-04-14 724d678
Docs requirements updated (#3)
