### 2021-04-14 8d02a86
Docs requirements updated (#3)
