### 2021-04-14 d5fc9ba
Docs requirements updated (#3)
