### 2021-04-14 80906f0
Docs requirements updated (#3)
