+++ README.rst
@@ -13,8 +13,9 @@ eemont
 .. image:: https://img.shields.io/badge/License-MIT-blue.svg
         :target: https://opensource.org/licenses/MIT
         
-.. image:: https://img.shields.io/badge/Documentation-0.1.9-green.svg
-        :target: https://eemont.readthedocs.io/en/0.1.9/
+.. image:: https://readthedocs.org/projects/eemont/badge/?version=latest
+        :target: https://eemont.readthedocs.io/en/latest/?badge=latest
+        :alt: Documentation Status
 
 .. image:: https://img.shields.io/badge/Donate-buy%20me%20a%20coffee-yellow.svg
         :target: https://www.buymeacoffee.com/davemlz
