### 2021-04-15 612361a
CONTRIBUTING added (#4, #5)
