### 2021-04-15 davemlz/eemont#5 @davemlz
Hi, @elbeejay!

I have added, for now, a general section for Contribution Guidelines that covers #4 and #5. You can find them in the [README](https://github.com/davemlz/eemont/blob/master/README.rst) and in the [CONTRIBUTING](https://github.com/davemlz/eemont/blob/master/CONTRIBUTING.rst) files. I plan to extend them in the future with more details on how to contribute for specific features of the package.

Thank you!
