+++ eemont/imagecollection.py
@@ -15,6 +15,10 @@ def _extend_eeImageCollection():
 def closest(self, date, tolerance = 1, unit = 'month'):
     '''Gets the closest image (or set of images if the collection intersects a region that requires multiple scenes) to the specified date.
     
+    Tip
+    ----------    
+    Check more info about getting the closest image to a specific date in the :ref:`User Guide<Closest Image to a Specific Date>`.
+    
     Parameters
     ----------    
     self : ee.ImageCollection [this]
@@ -66,6 +70,10 @@ def getTimeSeriesByRegion(self,reducer,bands = None,geometry = None,scale = None
                           maxPixels = 1e12,tileScale = 1,dateColumn = 'date',dateFormat = 'ISO',naValue = -9999):
     '''Gets the time series by region for the given image collection and geometry (feature or feature collection are also supported) according to the specified reducer (or reducers).
     
+    Tip
+    ----------    
+    Check more info about time series in the :ref:`User Guide<Time Series By Regions>`.
+    
     Parameters
     ----------
     self : ee.ImageCollection (this)
@@ -102,6 +110,10 @@ def getTimeSeriesByRegion(self,reducer,bands = None,geometry = None,scale = None
     ee.FeatureCollection
         Time series by region retrieved as a Feature Collection.
         
+    See Also
+    --------
+    getTimeSeriesByRegions : Gets the time series by regions for the given image collection and feature collection according to the specified reducer (or reducers).
+        
     Examples
     --------
     >>> import ee, eemont
@@ -119,10 +131,6 @@ def getTimeSeriesByRegion(self,reducer,bands = None,geometry = None,scale = None
     ...                               geometry = fc,
     ...                               bands = ['EVI','NDVI'],
     ...                               scale = 10)
-    
-    See Also
-    --------
-    getTimeSeriesByRegions : Gets the time series by regions for the given image collection and feature collection according to the specified reducer (or reducers).
     '''    
     if bands != None:
         if not isinstance(bands,list):
@@ -168,6 +176,10 @@ def getTimeSeriesByRegion(self,reducer,bands = None,geometry = None,scale = None
 def getTimeSeriesByRegions(self,reducer,collection,bands = None,scale = None,crs = None,crsTransform = None,tileScale = 1,dateColumn = 'date',dateFormat = 'ISO',naValue = -9999):
     '''Gets the time series by regions for the given image collection and feature collection according to the specified reducer (or reducers).
     
+    Tip
+    ----------    
+    Check more info about time series in the :ref:`User Guide<Time Series By Regions>`.
+    
     Parameters
     ----------
     self : ee.ImageCollection (this)
@@ -199,6 +211,11 @@ def getTimeSeriesByRegions(self,reducer,collection,bands = None,scale = None,crs
     ee.FeatureCollection
         Time series by regions retrieved as a Feature Collection.
         
+    See Also
+    --------
+    getTimeSeriesByRegion : Gets the time series by region for the given image collection and geometry (feature or feature collection are also supported)
+        according to the specified reducer (or reducers).
+        
     Examples
     --------
     >>> import ee, eemont
@@ -216,11 +233,6 @@ def getTimeSeriesByRegions(self,reducer,collection,bands = None,scale = None,crs
     ...                                collection = fc,
     ...                                bands = ['EVI','NDVI'],
     ...                                scale = 10)
-    
[…]
