+++ docs/classes/stubs/eemont.common.listIndices.rst
@@ -1,0 +1,6 @@
+eemont.common.listIndices
+=========================
+
+.. currentmodule:: eemont.common
+
+.. autofunction:: listIndices
