+++ docs/classes/eeimage.rst
@@ -1,5 +1,5 @@
-Image
-========
+ee.Image
+==========
 
 Extended methods for the ee.Image class:
 
