+++ docs/classes/eefeature.rst
@@ -1,4 +1,4 @@
-Feature
+ee.Feature
 ===========
 
 Extended methods for the ee.Feature class:
