+++ eemont/image.py
@@ -3,6 +3,7 @@ import warnings
 from .common import _get_platform
 from .common import _index
 from .common import _scale
+from .common import _maskClouds
 
 def _extend_eeImage():
     """Decorator. Extends the ee.ImageCollection class."""
@@ -727,201 +728,7 @@ def maskClouds(self, method = 'cloud_prob', prob = 60, maskCirrus = True, maskSh
     >>> ee.Initialize()
     >>> S2 = ee.ImageCollection('COPERNICUS/S2_SR').first().maskClouds(prob = 75,buffer = 300,cdi = -0.5)
     '''
-    def S3(args):
-        qa = args.select('quality_flags')
-        notCloud = qa.bitwiseAnd(1 << 27).eq(0);
-        return args.updateMask(notCloud)
-    
-    def S2(args):    
-
-        def cloud_prob(img):
-            clouds = ee.Image(img.get('cloud_mask')).select('probability')
-            isCloud = clouds.gte(prob).rename('CLOUD_MASK')
-            return img.addBands(isCloud)
-
-        def QA(img):
-            qa = img.select('QA60')
-            cloudBitMask = 1 << 10    
-            isCloud = qa.bitwiseAnd(cloudBitMask).eq(0)
-            if maskCirrus:
-                cirrusBitMask = 1 << 11
-                isCloud = isCloud.And(qa.bitwiseAnd(cirrusBitMask).eq(0))
-            isCloud = isCloud.Not().rename('CLOUD_MASK')
-            return img.addBands(isCloud)
-
-        def CDI(img):
-            idx = img.get('system:index')
-            S2TOA = ee.ImageCollection('COPERNICUS/S2').filter(ee.Filter.eq('system:index',idx)).first()
-            CloudDisplacementIndex = ee.Algorithms.Sentinel2.CDI(S2TOA)
-            isCloud = CloudDisplacementIndex.lt(cdi).rename('CLOUD_MASK_CDI')
-            return img.addBands(isCloud)
-
-        def get_shadows(img):      
-            notWater = img.select('SCL').neq(6)
-            if not scaledImage:
-                darkPixels = img.select('B8').lt(dark * 1e4).multiply(notWater)
-            else:  
-                darkPixels = img.select('B8').lt(dark).multiply(notWater)        
-            shadowAzimuth = ee.Number(90).subtract(ee.Number(img.get('MEAN_SOLAR_AZIMUTH_ANGLE')))        
-            cloudProjection = img.select('CLOUD_MASK').directionalDistanceTransform(shadowAzimuth,cloudDist/10)
-            cloudProjection = cloudProjection.reproject(crs = img.select(0).projection(),scale = 10).select('distance').mask()            
-            isShadow = cloudProjection.multiply(darkPixels).rename('SHADOW_MASK')        
-            return img.addBands(isShadow)
-
-        def clean_dilate(img):        
-            isCloudShadow = img.select('CLOUD_MASK')
-            if cdi != None:
-                isCloudShadow = isCloudShadow.And(img.select('CLOUD_MASK_CDI'))
-            if maskShadows:
-                isCloudShadow = isCloudShadow.add(img.select('SHADOW_MASK')).gt(0)        
-            isCloudShadow = isCloudShadow.focal_min(20,units = 'meters').focal_max(buffer*2/10,units = 'meters').rename('CLOUD_SHADOW_MASK')
[…]
