+++ eemont/common.py
@@ -979,3 +979,257 @@ def _scale(self):
         scaled = self.map(lookup[platformDict['platform']])
         
     return scaled
+
+def _maskClouds(self,method,prob,maskCirrus,maskShadows,scaledImage,dark,cloudDist,buffer,cdi):
+    '''Masks clouds and shadows in an image or image collection (valid just for Surface Reflectance products).
+    
+    Parameters
+    ----------    
+    self : ee.Image | ee.ImageCollection
+        Image or image collection to mask.
+    method : string, default = 'cloud_prob'
+        Method used to mask clouds.\n
+        Available options:
+            - 'cloud_prob' : Use cloud probability.
+            - 'qa' : Use Quality Assessment band.
+        This parameter is ignored for Landsat products.
+    prob : numeric [0, 100], default = 60
+        Cloud probability threshold. Valid just for method = 'prob'. This parameter is ignored for Landsat products.
+    maskCirrus : boolean, default = True
+        Whether to mask cirrus clouds. Valid just for method = 'qa'. This parameter is ignored for Landsat products.
+    maskShadows : boolean, default = True
+        Whether to mask cloud shadows. For more info see 'Braaten, J. 2020. Sentinel-2 Cloud Masking with s2cloudless. Google Earth Engine, Community Tutorials'.
+    scaledImage : boolean, default = False
+        Whether the pixel values are scaled to the range [0,1] (reflectance values). This parameter is ignored for Landsat products.
+    dark : float [0,1], default = 0.15
+        NIR threshold. NIR values below this threshold are potential cloud shadows. This parameter is ignored for Landsat products.
+    cloudDist : int, default = 1000
+        Maximum distance in meters (m) to look for cloud shadows from cloud edges. This parameter is ignored for Landsat products.
+    buffer : int, default = 250
+        Distance in meters (m) to dilate cloud and cloud shadows objects. This parameter is ignored for Landsat products.
+    cdi : float [-1,1], default = None
+        Cloud Displacement Index threshold. Values below this threshold are considered potential clouds. 
+        A cdi = None means that the index is not used. For more info see 'Frantz, D., HaS, E., Uhl, A., Stoffels, J., Hill, J. 2018. Improvement of the Fmask algorithm for Sentinel-2 images:
+        Separating clouds from bright surfaces based on parallax effects. Remote Sensing of Environment 2015: 471-481'.
+        This parameter is ignored for Landsat products.
+        
+    Returns
+    -------
+    ee.Image | ee.ImageCollection
+        Cloud-shadow masked image or image collection.
+    '''
+    def S3(args):
+        qa = args.select('quality_flags')
+        notCloud = qa.bitwiseAnd(1 << 27).eq(0);
+        return args.updateMask(notCloud)
+    
+    def S2(args):    
+
+        def cloud_prob(img):
+            clouds = ee.Image(img.get('cloud_mask')).select('probability')
+            isCloud = clouds.gte(prob).rename('CLOUD_MASK')
+            return img.addBands(isCloud)
+
+        def QA(img):
+            qa = img.select('QA60')
+            cloudBitMask = 1 << 10    
+            isCloud = qa.bitwiseAnd(cloudBitMask).eq(0)
[…]
