### 2021-04-18 a9bd55f
spectralIndices and scaleAndOffset added (#10)
