+++ eemont/image.py
@@ -551,7 +551,7 @@ def index(self,index = 'NDVI',G = 2.5,C1 = 6.0,C2 = 7.5,L = 1.0,kernel = 'RBF',s
     
     Warning
     -------------    
-    **Deprecation**: The :code:`index()` method will no longer be available for future versions. Please use :code:`spectralIndices()` instead.
+    **Pending Deprecation:** The :code:`index()` method will no longer be available for future versions. Please use :code:`spectralIndices()` instead.
     
     Tip
     ----------    
@@ -671,7 +671,134 @@ def index(self,index = 'NDVI',G = 2.5,C1 = 6.0,C2 = 7.5,L = 1.0,kernel = 'RBF',s
     - Computing all indices:
     
     >>> S2.index('all')
-    '''    
+    '''
+    warnings.warn("index() will be deprecated in future versions, please use spectralIndices() instead",PendingDeprecationWarning)
+    
+    return _index(self,index,G,C1,C2,L,kernel,sigma,p,c)
+
+@_extend_eeImage()
+def spectralIndices(self,index = 'NDVI',G = 2.5,C1 = 6.0,C2 = 7.5,L = 1.0,kernel = 'RBF',sigma = '0.5 * (a + b)',p = 2.0,c = 1.0):
+    '''Computes one or more spectral indices (indices are added as bands) for an image.
+        
+    Tip
+    ----------    
+    Check more info about the supported platforms and spectral indices in the :ref:`User Guide<Spectral Indices>`.
+    
+    Parameters
+    ----------    
+    self : ee.Image [this]
+        Image to compute indices on. Must be scaled to [0,1].      
+    index : string | list[string], default = 'NDVI'
+        Index or list of indices to compute.\n
+        Available options:
+            - 'vegetation' : Compute all vegetation indices.
+            - 'burn' : Compute all burn indices.
+            - 'water' : Compute all water indices.
+            - 'snow' : Compute all snow indices.
+            - 'drought' : Compute all drought indices.
+            - 'kernel' : Compute all kernel indices.
+            - 'all' : Compute all indices listed below.
+        Vegetation indices:
+            - 'BNDVI' : Blue Normalized Difference Vegetation Index.
+            - 'CIG' : Chlorophyll Index - Green.
+            - 'CVI' : Chlorophyll Vegetation Index.
+            - 'EVI' : Enhanced Vegetation Index.
+            - 'EVI2' : Two-Band Enhanced Vegetation Index.
+            - 'GARI' : Green Atmospherically Resistant Vegetation Index.
+            - 'GBNDVI' : Green-Blue Normalized Difference Vegetation Index.
+            - 'GEMI' : Global Environment Monitoring Index.
+            - 'GLI' : Green Leaf Index.
+            - 'GNDVI' : Green Normalized Difference Vegetation Index.
+            - 'GRNDVI' : Green-Red Normalized Difference Vegetation Index.
+            - 'GVMI' : Global Vegetation Moisture Index.
+            - 'MNDVI' : Modified Normalized Difference Vegetation Index.
+            - 'NDVI' : Normalized Difference Vegetation Index.
+            - 'NGRDI' : Normalized Green Red Difference Index.
+            - 'RVI' : Ratio Vegetation Index.
+            - 'SAVI' : Soil-Adjusted Vegetation Index.
+            - 'VARI' : Visible Atmospherically Resistant Index.
+        Burn and fire indices:     
+            - 'BAI' : Burned Area Index.
+            - 'BAIS2' : Burned Area Index for Sentinel 2.
+            - 'CSIT' : Char Soil Index Thermal.
+            - 'NBR' : Normalized Burn Ratio.
+            - 'NBRT' : Normalized Burn Ratio Thermal.
[…]
