### 2021-04-18 5798540
Tests for spectralIndices and scaleAndOffset (#10)
