+++ tests/test_image.py
@@ -87,221 +87,221 @@ class Test(unittest.TestCase):
     
     def test_S3(self):
         """Test the image module for Sentinel-3"""
-        S3_tested = S3.maskClouds().scale()
+        S3_tested = S3.maskClouds().scaleAndOffset()
         self.assertIsInstance(S3_tested, ee.image.Image)
     
     def test_S2(self):
         """Test the image module for Sentinel-2"""
-        S2_tested = S2.maskClouds().scale().index(indices)
+        S2_tested = S2.maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(S2_tested, ee.image.Image)
         
     # LANDSAT MISSIONS
         
     def test_L8(self):
         """Test the image module for Landsat 8"""
-        L8_tested = L8.maskClouds().scale().index(indices)
+        L8_tested = L8.maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L8_tested, ee.image.Image)
         
     def test_L7(self):
         """Test the image module for Landsat 7"""
-        L7_tested = L7.maskClouds().scale().index(indices)
+        L7_tested = L7.maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L7_tested, ee.image.Image)
         
     def test_L5(self):
         """Test the image module for Landsat 5"""
-        L5_tested = L5.maskClouds().scale().index(indices)
+        L5_tested = L5.maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L5_tested, ee.image.Image)
                 
     def test_L4(self):
         """Test the image module for Landsat 4"""
-        L4_tested = L4.maskClouds().scale().index(indices)
+        L4_tested = L4.maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L4_tested, ee.image.Image)
     
     # MODIS (TERRA + AQUA)
     
     def test_MCD43A4(self):
         """Test the image module for MCD43A4"""
-        MCD43A4_tested = MCD43A4.scale().index(indices)
+        MCD43A4_tested = MCD43A4.scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(MCD43A4_tested, ee.image.Image)
         
     def test_MCD43A3(self):
         """Test the image module for MCD43A3"""
-        MCD43A3_tested = MCD43A3.scale()
+        MCD43A3_tested = MCD43A3.scaleAndOffset()
         self.assertIsInstance(MCD43A3_tested, ee.image.Image)
         
     def test_MCD43A1(self):
         """Test the image module for MCD43A1"""
-        MCD43A1_tested = MCD43A1.scale()
+        MCD43A1_tested = MCD43A1.scaleAndOffset()
         self.assertIsInstance(MCD43A1_tested, ee.image.Image)
         
     def test_MCD15A3H(self):
         """Test the image module for MCD15A3H"""
-        MCD15A3H_tested = MCD15A3H.maskClouds().scale()
+        MCD15A3H_tested = MCD15A3H.maskClouds().scaleAndOffset()
         self.assertIsInstance(MCD15A3H_tested, ee.image.Image)
         
     # MODIS (TERRA)
         
     def test_MOD09GQ(self):
         """Test the image module for MOD09GQ"""
-        MOD09GQ_tested = MOD09GQ.scale().index(indices)
+        MOD09GQ_tested = MOD09GQ.scaleAndOffset().spectralIndices(indices)
[…]
