+++ tests/test_imagecollection.py
@@ -67,221 +67,221 @@ class Test(unittest.TestCase):
     
     def test_S3(self):
         """Test the image collection module for Sentinel-3"""
-        S3_tested = S3.closest('2020-01-15').maskClouds().scale()
+        S3_tested = S3.closest('2020-01-15').maskClouds().scaleAndOffset()
         self.assertIsInstance(S3_tested, ee.imagecollection.ImageCollection)
     
     def test_S2(self):
         """Test the image collection module for Sentinel-2"""
-        S2_tested = S2.closest('2020-01-15').maskClouds().scale().index(indices)
+        S2_tested = S2.closest('2020-01-15').maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(S2_tested, ee.imagecollection.ImageCollection)
     
     # LANDSAT MISSIONS
     
     def test_L8(self):
         """Test the image collection module for Landsat 8"""
-        L8_tested = L8.closest('2020-01-15').maskClouds().scale().index(indices)
+        L8_tested = L8.closest('2020-01-15').maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L8_tested, ee.imagecollection.ImageCollection)
         
     def test_L7(self):
         """Test the image collection module for Landsat 7"""
-        L7_tested = L7.closest('2010-01-15').maskClouds().scale().index(indices)
+        L7_tested = L7.closest('2010-01-15').maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L7_tested, ee.imagecollection.ImageCollection)
         
     def test_L5(self):
         """Test the image collection module for Landsat 5"""
-        L5_tested = L5.closest('2000-01-15',2,'year').maskClouds().scale().index(indices)
+        L5_tested = L5.closest('2000-01-15',2,'year').maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L5_tested, ee.imagecollection.ImageCollection)
         
     def test_L4(self):
         """Test the image collection module for Landsat 4"""
-        L4_tested = L4.closest('1990-01-15').maskClouds().scale().index(indices)
+        L4_tested = L4.closest('1990-01-15').maskClouds().scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(L4_tested, ee.imagecollection.ImageCollection)
     
     # MODIS (TERRA + AQUA)
     
     def test_MCD43A4(self):
         """Test the image collection module for MCD43A4"""
-        MCD43A4_tested = MCD43A4.closest('2020-01-15').scale().index(indices)
+        MCD43A4_tested = MCD43A4.closest('2020-01-15').scaleAndOffset().spectralIndices(indices)
         self.assertIsInstance(MCD43A4_tested, ee.imagecollection.ImageCollection)
         
     def test_MCD43A3(self):
         """Test the image collection module for MCD43A3"""
-        MCD43A3_tested = MCD43A3.closest('2020-01-15').scale()
+        MCD43A3_tested = MCD43A3.closest('2020-01-15').scaleAndOffset()
         self.assertIsInstance(MCD43A3_tested, ee.imagecollection.ImageCollection)
         
     def test_MCD43A1(self):
         """Test the image collection module for MCD43A1"""
-        MCD43A1_tested = MCD43A1.closest('2020-01-15').scale()
+        MCD43A1_tested = MCD43A1.closest('2020-01-15').scaleAndOffset()
[…]
