### 2021-04-21 95890f3
Updated GEE STAC IDs (#12)
