+++ eemont/imagecollection.py
@@ -1,7 +1,6 @@
 import ee
 import numpy as np
 import warnings
-from .common import _get_platform
 from .common import _index
 from .common import _scale
 from .common import _maskClouds
