+++ eemont/common.py
@@ -1,6 +1,142 @@
 import ee
+import json
+import pkg_resources
+import os
 import warnings
 from box import Box
+
+# Platforms IDs
+# --------------------------
+
+def _get_platform(args):
+    '''Gets the platform (satellite) of an image (or image collection) and wheter if it is a Surface Reflectance product.
+    
+    Parameters
+    ----------
+    args : ee.Image | ee.ImageCollection
+        Image to get platform.
+        
+    Returns
+    -------
+    dict
+        Platform and product of the image (or image collection).
+    '''
+    platforms = [
+        'COPERNICUS/S3',
+        'COPERNICUS/S2',
+        'LANDSAT/LC08',
+        'LANDSAT/LE07',
+        'LANDSAT/LT05',
+        'LANDSAT/LT04',
+        'MODIS/006/MCD43A4',
+        'MODIS/006/MCD43A3',
+        'MODIS/006/MOD09GQ',
+        'MODIS/006/MOD10A1',
+        'MODIS/006/MOD11A1',
+        'MODIS/006/MOD09GA',
+        'MODIS/006/MODOCGA',
+        'MODIS/006/MOD14A1',
+        'MODIS/006/MCD43A1',
+        'MODIS/006/MCD15A3H',
+        'MODIS/006/MOD09Q1',
+        'MODIS/006/MOD09A1',
+        'MODIS/006/MOD11A2',
+        'MODIS/006/MOD17A2H',
+        'MODIS/006/MOD16A2',
+        'MODIS/006/MOD13Q1',
+        'MODIS/006/MOD13A1',
+        'MODIS/006/MOD13A2',
+        'MODIS/061/MOD08_M3',
+        'MODIS/006/MOD17A3HGF',
+        'MODIS/006/MYD09GQ',
+        'MODIS/006/MYD10A1',
+        'MODIS/006/MYD11A1',
+        'MODIS/006/MYD09GA',
+        'MODIS/006/MYDOCGA',
+        'MODIS/006/MYD14A1',        
+        'MODIS/006/MYD09Q1',
+        'MODIS/006/MYD09A1',
+        'MODIS/006/MYD11A2',
+        'MODIS/006/MYD17A2H',        
+        'MODIS/006/MYD13Q1',
+        'MODIS/006/MYD13A1',
+        'MODIS/006/MYD13A2',
+        'MODIS/061/MYD08_M3',
+        'MODIS/006/MYD17A3HGF',
+        'NOAA/VIIRS/001/VNP09GA',
+        'NOAA/VIIRS/001/VNP13A1'
+    ]
+    
+    if isinstance(args, ee.imagecollection.ImageCollection):
+        imgID = args.first().get('system:id').getInfo()
+    elif isinstance(args, ee.image.Image):
+        imgID = args.get('system:id').getInfo()
+    else:
+        raise Exception("Sorry, just ee.Image or ee.ImageCollection classes supported!")
+    
+    plt = None
+    
+    for platform in platforms:
+        if platform in imgID:
+            plt = platform
+        if '_SR' in imgID:
+            platformDict = {'platform': plt, 'sr': True}
+        else:
+            platformDict = {'platform': plt, 'sr': False}
+            
+    if plt is None:
+        raise Exception("Sorry, satellite platform not supported!")
+            
+    return platformDict
+
+def _get_platform_STAC(args):
+    '''Gets the platform (satellite) of an image (or image collection) and wheter if it is a Surface Reflectance product.
+    
+    Parameters
+    ----------
+    args : ee.Image | ee.ImageCollection
+        Image or image collection to get the platform from.
[…]
