### 2021-04-21 58d2a04
scale and offset from the GEE STAC (#13)
