### 2021-04-21 e720ca0
Overloaded operators for ee.Number (#6)
