+++ eemont/number.py
@@ -1,0 +1,542 @@
+import ee
+import warnings
+
+def _extend_eeNumber():
+    """Decorator. Extends the ee.Number class."""
+    return lambda f: (setattr(ee.ee_number.Number,f.__name__,f) or f)
+
+@_extend_eeNumber()
+def __add__(self, other):
+    '''Computes the addition between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Left operand.
+    other : ee.Number | numeric 
+        Right operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Addition of two numbers.
+    '''
+    return self.add(other)
+
+@_extend_eeNumber()
+def __radd__(self, other):
+    '''Computes the addition between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Right operand.
+    other : ee.Number | numeric 
+        Left operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Addition of two numbers.
+    '''
+    return self.add(other)
+
+@_extend_eeNumber()
+def __sub__(self, other):
+    '''Computes the subtraction between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Left operand.
+    other : ee.Number | numeric 
+        Right operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Subtraction of two numbers.
+    '''
+    return self.subtract(other)
+
+@_extend_eeNumber()
+def __rsub__(self, other):
+    '''Computes the subtraction between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Right operand.
+    other : ee.Number | numeric 
+        Left operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Subtraction of two numbers.
+    '''
+    return ee.Number(other).subtract(self)
+
+@_extend_eeNumber()
+def __mul__(self, other):
+    '''Computes the multiplication between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Left operand.
+    other : ee.Number | numeric 
+        Right operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Multiplication of two numbers.
+    '''
+    return self.multiply(other)
+
+@_extend_eeNumber()
+def __rmul__(self, other):
+    '''Computes the multiplication between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Right operand.
+    other : ee.Number | numeric 
+        Left operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Multiplication of two numbers.
+    '''
+    return self.multiply(other)
+
+@_extend_eeNumber()
+def __truediv__(self, other):
+    '''Computes the division between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Left operand.
+    other : ee.Number | numeric 
+        Right operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Division of two numbers.
+    '''
+    return self.divide(other)
+
+@_extend_eeNumber()
+def __rtruediv__(self, other):
+    '''Computes the division between two numbers.
+
+    Parameters
+    ----------
+    self : ee.Number
+        Right operand.
+    other : ee.Number | numeric 
+        Left operand. If numeric, an ee.Number is created from its value.
+
+    Returns
+    -------
+    ee.Number
+        Division of two numbers.
[…]
