+++ eemont/__init__.py
@@ -1,6 +1,6 @@
 """The eemont python package."""
 
-__version__ = '0.1.9'
+__version__ = "0.1.9"
 
 from . import feature
 from . import featurecollection
