+++ eemont/geometry.py
@@ -2,20 +2,22 @@ import ee
 import geopy
 from geopy.geocoders import get_geocoder_for_service
 
+
 def _extend_staticmethod_eeGeometry():
     """Decorator. Extends the ee.Geometry class with a static method."""
-    return lambda f: (setattr(ee.geometry.Geometry,f.__name__,staticmethod(f)) or f)
+    return lambda f: (setattr(ee.geometry.Geometry, f.__name__, staticmethod(f)) or f)
+
 
 @_extend_staticmethod_eeGeometry()
-def BBoxFromQuery(query,geocoder = "nominatim", **kwargs):
-    '''Constructs an ee.Geometry describing a bounding box from a query submitted to a geodocer using the geopy package.
-    
+def BBoxFromQuery(query, geocoder="nominatim", **kwargs):
+    """Constructs an ee.Geometry describing a bounding box from a query submitted to a geodocer using the geopy package.
+
     Tip
-    ----------    
+    ----------
     Check more info about constructors in the :ref:`User Guide<Constructors>`.
-    
+
     Parameters
-    ----------    
+    ----------
     query : str
         Address, query or structured query to geocode.
     geocoder : str, default = 'nominatim'
@@ -23,17 +25,17 @@ def BBoxFromQuery(query,geocoder = "nominatim", **kwargs):
     **kwargs :
         Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
         user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
-        
+
     Returns
     -------
     ee.Geometry.Polygon
         Geometry describing a bounding box from the specified query.
-    
+
     See Also
     --------
     PointFromQuery : Constructs an ee.Geometry describing a point from a query submitted to a geodocer using the geopy package.
     MultiPointFromQuery : Constructs an ee.Geometry describing a multi-point from a query submitted to a geodocer using the geopy package.
-    
+
     Examples
     --------
     >>> import ee, eemont
@@ -72,35 +74,40 @@ def BBoxFromQuery(query,geocoder = "nominatim", **kwargs):
         }
       }
     })
-    '''
-    if geocoder in ['nominatim','arcgis']:
+    """
+    if geocoder in ["nominatim", "arcgis"]:
         cls = get_geocoder_for_service(geocoder)
     else:
         raise Exception('Invalid geocoder! Use one of "nominatim" or "arcgis".')
     geolocator = cls(**kwargs)
     location = geolocator.geocode(query)
     if location is None:
-        raise Exception('No matches were found for your query!')
+        raise Exception("No matches were found for your query!")
     else:
-        if geocoder == 'nominatim':
-            BBox = location.raw['boundingbox']
-            return ee.Geometry.BBox(float(BBox[2]),float(BBox[0]),float(BBox[3]),float(BBox[1]))
-        elif geocoder == 'arcgis':
-            BBox = location.raw['extent']
-            return ee.Geometry.BBox(BBox['xmin'],BBox['ymin'],BBox['xmax'],BBox['ymax'])
+        if geocoder == "nominatim":
+            BBox = location.raw["boundingbox"]
+            return ee.Geometry.BBox(
+                float(BBox[2]), float(BBox[0]), float(BBox[3]), float(BBox[1])
+            )
+        elif geocoder == "arcgis":
+            BBox = location.raw["extent"]
+            return ee.Geometry.BBox(
+                BBox["xmin"], BBox["ymin"], BBox["xmax"], BBox["ymax"]
+            )
         else:
[…]
