+++ eemont/__init__.py
@@ -9,4 +9,5 @@ from . import image
 from . import imagecollection
 from . import dataframe
 from . import number
+from . import eeList
 from .common import *
