### 2021-04-26 davemlz/eemont#18 @elbeejay
[JOSS Review] Errors in tutorial notebooks
Hello, I was running through the tutorial notebooks (locally) and ran into the following errors:

`004-Computing-Spectral-Indices-Landsat-8.ipynb`, I received the following error in the final cell:
```
<ipython-input-16-e652de6f3203> in <module>
      2 Map.addLayer(L8.select('GNDVI'),visParamsGNDVI,'GNDVI')
      3 Map.addLayer(L8.select('NDWI'),visParamsNDWI,'NDWI')
----> 4 Map.add_colorbar(visParamsGNDVI['palette'], caption = 'GNDVI')
      5 Map.add_colorbar(visParamsNDWI['palette'], caption = 'NDWI')
      6 Map.addLayerControl()

~/anaconda3/envs/eemont/lib/python3.9/site-packages/geemap/geemap.py in add_colorbar(self, vis_params, cmap, discrete, label, orientation, position, transparent_bg, layer_name, **kwargs)
   2605
   2606         if not isinstance(vis_params, dict):
-> 2607             raise TypeError("The vis_params must be a dictionary.")
   2608
   2609         if orientation not in ["horizontal", "vertical"]:

TypeError: The vis_params must be a dictionary.
```
I had similar errors in tutorials 005 and 006 as well.

---

Then in `013-Time-Series-By-Region-Pandas.ipynb`, I had the following code-error combo.
Code block:
```
tsPandas = pd.melt(tsPandas,
                   id_vars = ['system:index','reducer','date'],
                   value_vars = ['GNDVI','EVI'],
                   var_name = 'Index',
                   value_name = 'Value')
```
Error:
```
KeyError                                  Traceback (most recent call last)
<ipython-input-11-bd846c0b8d50> in <module>
----> 1 tsPandas = pd.melt(tsPandas,
      2                    id_vars = ['system:index','reducer','date'],
      3                    value_vars = ['GNDVI','EVI'],
      4                    var_name = 'Index',
      5                    value_name = 'Value')

~/anaconda3/envs/eemont/lib/python3.9/site-packages/pandas/core/reshape/melt.py in melt(frame, id_vars, value_vars, var_name, value_name, col_level, ignore_index)
     62             missing = Index(com.flatten(id_vars)).difference(cols)
     63             if not missing.empty:
---> 64                 raise KeyError(
     65                     "The following 'id_vars' are not present "
     66                     f"in the DataFrame: {list(missing)}"

KeyError: "The following 'id_vars' are not present in the DataFrame: ['system:index']"
```
I ran into a similar error in tutorial 014 as well.

---

I had no issues running tutorials 001, 002, or 003. For tutorials 007 and 008, I wasn't able to get the maps to load, but this may be an error (or impatience) on my end, as the "Loading widget..." text did come up.

This issue relates to the ongoing [JOSS Review](https://github.com/openjournals/joss-reviews/issues/3168) of this package.
