### 2021-04-26 davemlz/eemont#18 @davemlz
Hi, @elbeejay!

The error of the tutorials 004, 005 and 006 is related to how `geemap` is imported. Since the tutorials were created to work in Google Colab, `geemap` methods are related to `folium` and not to `ipyleaflet`. When you run the notebooks locally, `geemap` uses `ipyleaflet` by default, this can be modified by importing `geemap` as follows:

```python
import geemap.eefolium as geemap
```

That should work for those tutorials. (I will add a comment indicating this in the tutorials, thank you for pointing this out!)

In the case of tutorials 013 and 014, this error is related to the EarthEngine API since it is not longer returning the `system:index` property anymore. I have to modify these two tutorials.

I will modify the tutorials and I'll let you know!

Thank you very much!
