### 2021-04-26 af71880
Tutorials modified (#18)
