### 2021-04-26 davemlz/eemont#18 @davemlz
Hi, @elbeejay!

I have modified the tutorials.

Let me know if they work!

Cheers!
