### 2021-04-26 davemlz/eemont#19 @elbeejay
[JOSS Review] Paper/documentation comments and suggestions
The `eemont` package is clearly a useful addition to the GEE API and improves the functionality and ease of use of Google Earth Engine, so congratulations to @davemlz for creating a useful package for the broader gee community.

That being said I think both the documentation and the JOSS paper could make a stronger case for the benefits of the `eemont` package. From the paper and documentation alone is not obvious to me why I would pick up `eemont` - potentially simpler code than the standard Earth Engine API? I would suggest showing the necessary code *without `eemont`* to perform the same analysis for one or two of the examples in the [Features](https://github.com/davemlz/eemont#features) section of the documentation alongside the `eemont` implementation. Having a direct comparison of the (presumably longer) standard Earth Engine code required to accomplish the same task you can perform with `eemont` in just a few lines provides a much more compelling reason for someone who sees the package to use it.

In a similar vein, in the context of the paper, I think it would at least be beneficial to state the number of lines required to perform the process in the example given with and without `eemont`. This way a reader of the JOSS paper will be able to more intuitively grasp the value of using `eemont`. Even better, if there is some estimated "average" simplification or line-shortening accomplished by using `eemont` that'd be great to include, although I can understand that such a quantification may be impossible due to the diversity of remote sensing applications and processes `eemont` can be used for.

Regarding the "State of the field", I believe there are a handful of packages out there that provide additional functionality on-top of the standard Earth Engine API (e.g., `geemap`, `geetools`, some packages by GitHub users fitoprincipe and samapriya, just to highlight a few). I think the paper should mention more of these other tools and packages, as `eemont` can likely improve users' experiences with them as well (at least where calls to Earth Engine are made). Furthermore it would provide a good place to highlight the unique niche `eemont` fills amongst other tools and packages that build off of the standard Earth Engine API.

---

This comment relates to the ongoing [JOSS review](https://github.com/openjournals/joss-reviews/issues/3168) of this package, comment pertains to the "Documentation" and "Software paper" sections of the JOSS reviewer guidelines:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

> A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

> State of the field: Do the authors describe how this software compares to other commonly-used packages?
