### 2021-04-26 #3168 @elbeejay
I've gone though the package and opened up a few issues ([#18](https://github.com/davemlz/eemont/issues/18) and [#19](https://github.com/davemlz/eemont/issues/19) in the `eemont` repository) where I outline my outstanding questions, concerns, and suggestions regarding this submission.

But overall I think the `eemont` is a valuable contribution to the Python+Google Earth Engine community, it seems to have picked up a reasonable user-base already and is featured on the [GEE developer resources page](https://developers.google.com/earth-engine/tutorials/community/developer-resources). Once those 2 outstanding issues are taken care of I will be able to complete the checklist and recommend the `eemont` software and JOSS paper for publication.
