### 2021-04-26 davemlz/eemont#19 @davemlz
Hi, @elbeejay!

Thank you very much for your comments and suggestions!

I really like your ideas for the documentation and also for the paper! I will work on them and I'll let you know when they're ready.

Thank you very much again!

Cheers!
