### 2021-04-26 davemlz/eemont#18 @elbeejay
The revised examples all worked for me with the exception of tutorial 012. I had to fiddle a bit with the import statement and colorbar statements to get the plots to show up properly.
