### 2021-04-27 davemlz/eemont#19 @davemlz
Hi, @elbeejay!

I have updated the examples of the features in the [README](https://github.com/davemlz/eemont/blob/master/README.rst) and also in the [Docs for the upcoming version](https://eemont.readthedocs.io/en/latest/?badge=latest). The new examples now show the comparison between using just the GEE Python API methods and using eemont methods.

Now I'm going to work on the paper with your additional suggestions!

Thank you!
