+++ README.rst
@@ -87,68 +87,228 @@ Install the latest eemont version from PyPI by running:
 Features
 --------
 
-The following features are extended through eemont:
+Let's see some of the main features of eemont and how simple they are compared to the GEE Python API original methods:
+
+Overloaded Operators
+~~~~~~~~~~~~~~~~~~~~~~~
+
+The following operators are overloaded: +, -, \*\, /, //, %, \**\ , <<, >>, &, \|\, <, <=, ==, !=, >, >=, -, ~. (and you can avoid the :code:`ee.Image.expression()` method!)
+
+.. list-table::
+   :widths: 50 50
+   :header-rows: 1
+
+   * - GEE Python API
+     - eemont-style     
+   * - .. code-block:: python             
+          
+          ds = 'COPERNICUS/S2_SR'
+          
+          S2 = (ee.ImageCollection(ds)
+            .first())
+            
+          exp = '2.5*(N-R)/(N+(6*R)-(7.5*B)+1)'
+          
+          imgDict = {
+            'N': S2.select('B8'),
+            'R': S2.select('B4'),
+            'B': S2.select('B2')
+          }
+   
+          EVI = S2.expression(exp,imgDict)
+     - .. code-block:: python                     
+   
+          ds = 'COPERNICUS/S2_SR'
+          
+          S2 = (ee.ImageCollection(ds)
+            .first())
+
+          N = S2.select('B8')
+          R = S2.select('B4')
+          B = S2.select('B2')
+
+          EVI = 2.5*(N-R)/(N+(6*R)-(7.5*B)+1)
+
+Clouds and Shadows Masking
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Masking clouds and shadows can be done using eemont with just one method: :code:`maskClouds()`!
+
+.. list-table::
+   :widths: 50 50
+   :header-rows: 1
+
+   * - GEE Python API
+     - eemont-style     
+   * - .. code-block:: python             
+          
+          ds = 'LANDSAT/LC08/C01/T1_SR'
+          
+          def maskCloudsShadows(img):
+              c = (1 << 3)
+              s = (1 << 5)
+              qa = 'pixel_qa'
+              qa = img.select(qa)
+              cm = qa.bitwiseAnd(c).eq(0)
+              sm = qa.bitwiseAnd(s).eq(0)
+              mask = cm.And(sm)
+              return img.updateMask(mask)
+              
+          (ee.ImageCollection(ds)
+            .map(maskCloudsShadows))
+     - .. code-block:: python                 
+   
+          ds = 'LANDSAT/LC08/C01/T1_SR'
+          
+          (ee.ImageCollection(ds)
+            .maskClouds())
+
+Image Scaling and Offsetting
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Scaling and offsetting can also be done using eemont with just one method: :code:`scale()`!
+
+.. list-table::
+   :widths: 50 50
+   :header-rows: 1
+
+   * - GEE Python API
+     - eemont-style     
+   * - .. code-block:: python                    
+   
+          def scaleBands(img):
+              scaling = img.select([
+                'NDVI',
+                'EVI',
+                'sur.*'
+              ])
+              x = scaling.multiply(0.0001)
+              scaling = img.select('.*th')
+              scaling = scaling.multiply(0.01)
+              x = x.addBands(scaling)
+              notScaling = img.select([
+                'DetailedQA',
+                'DayOfYear',
+                'SummaryQA'
+              ])
+              return x.addBands(notScaling)              
+          
+          ds = 'MODIS/006/MOD13Q1'
+          
[…]
