### 2021-04-28 26b9646
Awesome Spectral Indices added (#15)
