+++ eemont/common.py
@@ -179,298 +179,13 @@ def _get_indices():
     dict
         Indices.
     """
-    vegetationIndices = {
-        "BNDVI": {
-            "formula": "(N - B)/(N + B)",
-            "description": "Blue Normalized Difference Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "B"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=135",
-            "contributor": "davemlz",
-        },
-        "CIG": {
-            "formula": "(N / G) - 1.0",
-            "description": "Blue Normalized Difference Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "G"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=128",
-            "contributor": "davemlz",
-        },
-        "CVI": {
-            "formula": "(N * R) / (G ** 2.0)",
-            "description": "Chlorophyll Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "R", "G"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=391",
-            "contributor": "davemlz",
-        },
-        "EVI": {
-            "formula": "g * (N - R) / (N + C1 * R - C2 * B + L)",
-            "description": "Enhanced Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "g", "R", "C1", "C2", "B", "L"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=16",
-            "contributor": "davemlz",
-        },
-        "EVI2": {
-            "formula": "g * (N - R) / (N + 2.4 * R + L)",
-            "description": "Two-Band Enhanced Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "g", "R", "L"],
-            "reference": "https://doi.org/10.1016/j.rse.2008.06.006",
-            "contributor": "davemlz",
-        },
-        "GARI": {
-            "formula": "(N - (G - (B - R))) / (N - (G + (B - R)))",
-            "description": "Green Atmospherically Resistant Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "G", "B", "R"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=363",
-            "contributor": "davemlz",
-        },
-        "GBNDVI": {
-            "formula": "(N - (G + B))/(N + (G + B))",
-            "description": "Green-Blue Normalized Difference Vegetation Index",
-            "type": "vegetation",
-            "requires": ["N", "B", "G"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=186",
-            "contributor": "davemlz",
-        },
-        "GEMI": {
-            "formula": "((2.0*((N ** 2.0)-(R ** 2.0)) + 1.5*N + 0.5*R)/(N + R + 0.5))*(1.0 - 0.25*((2.0 * ((N ** 2.0) - (R ** 2)) + 1.5 * N + 0.5 * R)/(N + R + 0.5)))-((R - 0.125)/(1 - R))",
-            "description": "Global Environment Monitoring Index",
-            "type": "vegetation",
-            "requires": ["N", "R"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=25",
-            "contributor": "davemlz",
-        },
-        "GLI": {
-            "formula": "(2.0 * G - R - B) / (2.0 * G + R + B)",
-            "description": "Green Leaf Index",
-            "type": "vegetation",
-            "requires": ["G", "B", "R"],
-            "reference": "https://www.indexdatabase.de/db/i-single.php?id=375",
-            "contributor": "davemlz",
-        },
-        "GNDVI": {
[…]
