+++ .github/workflows/update_awesome_spectral_indices.yml
@@ -1,0 +1,37 @@
+name: update_awesome_spectral_indices
+on:
+  workflow_dispatch:
+  schedule:
+    - cron:  '0 0 * * *'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v2
+      - name: python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.7'          
+      - name: dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install requests
+      - name: execute        
+        run: |
+          python ./.github/scripts/update_awesome_spectral_indices.py
+      - name: commit
+        continue-on-error: true
+        run: |
+          today=$(date +"%Y-%m-%d %H:%M:%S")
+          git config --local user.email "action@github.com"
+          git config --local user.name "GitHub Action"
+          git add -A
+          git commit -m "Updated Awesome Spectral Indices ${today} UTC" -a
+      - name: push
+        continue-on-error: true
+        uses: ad-m/github-push-action@v0.6.0
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          branch: master
