+++ .github/workflows/tests.yml
@@ -1,0 +1,31 @@
+name: tests
+
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: [3.6, 3.7, 3.8, 3.9]
+        os: [ubuntu-latest, macOS-latest, windows-latest]
+    steps:
+      - name: checkout
+        uses: actions/checkout@v2
+      - name: python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade tox
+      - name: execute        
+        run: |
+          python ./.github/scripts/update_ee_token.py
+        env:
+          EE_TOKEN: ${{ secrets.EE_TOKEN }}
+      - name: tests
+        run: |
+          tox -e py
