+++ tox.ini
@@ -2,5 +2,5 @@
 envlist = py36,py37,py38,py39
 
 [testenv]
-commands = py.test eemont
+commands = pytest tests
 deps = pytest
