### 2021-04-29 davemlz/eemont#19 @elbeejay
Looks great, I think it makes the package very attractive for someone that stumbles upon it while trying to pick up the GEE Python API
