### 2021-04-29 davemlz/eemont#18 @giswqs
@davemlz I just improved the geemap `add_colorbar()` function (https://github.com/giswqs/geemap/commit/dc7e54856694a1994b6d4f4044385babe04bd086), making it compatible with both folium and ipyleaflet. It should be available in the next geemap release. You can also run `geemap.update_package()` to install geemap from GitHub. With the improved `add_colorbar()` function, your notebooks will no longer run into errors, and you can just use `import geemap` rather than `geemap.eefolium as geemap`. In short, you notebooks should work with both Colab and local Jupyter without making any changes.

![image](https://user-images.githubusercontent.com/5016453/116496189-7a81d800-a872-11eb-8b8f-caa62abca1b2.png)
