### 2021-04-29 davemlz/eemont#18 @giswqs
@davemlz I just released geemap [v0.8.15](https://github.com/giswqs/geemap/releases/tag/v0.8.15), which includes the improved `add_colorbar` function. Your notebooks will now work with both Colab and local Jupyter notebooks without making any changes. Also note that `Map.addLayerControl()` is not needed. geemap will automatically add the layer control when the map is displayed in Colab. `Map = geemap.Map()` automatically handles `ee.Authenticate()` and `ee.Initialize()`. Therefore, if you can write `Map = geemap.Map()` after `import geemap`, you will not need  `ee.Authenticate()` or `ee.Initialize()`

https://github.com/openjournals/joss-reviews/issues/3168
