### 2021-04-29 #3168 @giswqs
I just released geemap [v0.8.15](https://github.com/giswqs/geemap/releases/tag/v0.8.15), which should help resolve the issue raised in https://github.com/davemlz/eemont/issues/18 regarding the eemont tutorial notebooks.
