### 2021-04-29 davemlz/eemont#18 @davemlz
I have updated the tutorials (https://github.com/davemlz/eemont/commit/320c867413aa8176bc3abe7a531ebc1ee0d225be). Thank you @giswqs for the update in `geemap`! However, I had some trouble in [Tutorial 012](https://github.com/davemlz/eemont/blob/master/tutorials/012-Spectral-Indices-MODIS-MOD09GA.ipynb) when using the `colormaps`. I think it is caused by the return type of the `Box` object, converting it to a list solved the problem. I can open an Issue in the repository of `geemap` with the full code and errors.
