### 2021-04-30 davemlz/eemont#19 @davemlz
Hi, @elbeejay!

I have modified the paper following your suggestions. I have added the comparison with the GEE Python API, and also two new sections: one for the GEE Community Developer Resources and one for the integration with QGIS (both for the state of the field).

Cheers!
