### 2021-04-30 #3168 @davemlz
Hi to everyone!

Following @elbeejay suggestions (https://github.com/davemlz/eemont/issues/19), I have modified the paper (https://github.com/davemlz/eemont/commit/5e50b13a43ae90a5c8d0a0686ccb74934a426ece) with new sections (including the GEE Community: Developer Resources and the integration with the Earth Engine Plugin for QGIS), benefits of the package (including comparisons with the Earth Engine Python API), and I adjusted it for the State of the Field.
