+++ paper/paper.bib
@@ -38,3 +38,33 @@
   note = {R package version 1.18},
   url = {https://rstudio.github.io/reticulate/},
 }
+
+@Manual{Markert2021,
+  title = {{restee}: Python package to call processed EE objects via the REST API to local data},
+  author = {Kel Markert},
+  year = {2021},  
+  url = {https://kmarkert.github.io/restee/},
+}
+
+@Manual{Principe2021,
+  title = {{geetools}: Google Earth Engine Tools (scripts)},
+  author = {Rodrigo principe},
+  year = {2021},  
+  url = {https://github.com/gee-community/gee_tools},
+}
+
+@Manual{QGIS_software,
+  title = {QGIS Geographic Information System},
+  author = {{QGIS Development Team}},
+  organization = {QGIS Association},
+  year = {2021},
+  url = {https://www.qgis.org},
+}
+
+@Manual{Donchyts2021,
+  title = {Google Earth Engine Plugin for QGIS},
+  author = {Gennadii Donchyts},
+  year = {2021},
+  url = {https://gee-community.github.io/qgis-earthengine-plugin/},
+}
+
