+++ tests/test_common.py
@@ -13,10 +13,20 @@ class Test(unittest.TestCase):
         test = eemont.indices()
         self.assertIsInstance(test, box.box.Box)
         
+    def test_indices_online(self):
+        """Test the indices function"""
+        test = eemont.indices(online = True)
+        self.assertIsInstance(test, box.box.Box)
+        
     def test_listIndices(self):
         """Test the listIndices function"""
         test = eemont.listIndices()
-        self.assertIsInstance(test, list)        
+        self.assertIsInstance(test, list)  
+        
+    def test_listIndices_online(self):
+        """Test the listIndices function"""
+        test = eemont.listIndices(online = True)
+        self.assertIsInstance(test, list)
         
 if __name__ == '__main__':
     unittest.main()
