### 2021-05-01 751dd82
preprocess method (#16)
