+++ tests/test_image.py
@@ -313,6 +313,236 @@ class Test(unittest.TestCase):
         """Test the image module for MYD17A3HGF"""
         MYD17A3HGF_tested = MYD17A3HGF.scaleAndOffset()
         self.assertIsInstance(MYD17A3HGF_tested, ee.image.Image)
+
+    # ----------------------
+    # CHECKING PREPROCESS
+    # ----------------------
+        
+    # SENTINEL MISSIONS
+    
+    def test_S3_preprocess(self):
+        """Test the image module for Sentinel-3"""
+        S3_tested = S3.preprocess()
+        self.assertIsInstance(S3_tested, ee.image.Image)
+    
+    def test_S2_preprocess(self):
+        """Test the image module for Sentinel-2"""
+        S2_tested = S2.preprocess().spectralIndices(indices)
+        self.assertIsInstance(S2_tested, ee.image.Image)
+        
+    def test_S2_online_preprocess(self):
+        """Test the image module for Sentinel-2"""
+        S2_tested = S2.preprocess().spectralIndices(indices,online = True)
+        self.assertIsInstance(S2_tested, ee.image.Image)
+        
+    # LANDSAT MISSIONS
+        
+    def test_L8_preprocess(self):
+        """Test the image module for Landsat 8"""
+        L8_tested = L8.preprocess().spectralIndices(indices)
+        self.assertIsInstance(L8_tested, ee.image.Image)
+        
+    def test_L7_preprocess(self):
+        """Test the image module for Landsat 7"""
+        L7_tested = L7.preprocess().spectralIndices(indices)
+        self.assertIsInstance(L7_tested, ee.image.Image)
+        
+    def test_L5_preprocess(self):
+        """Test the image module for Landsat 5"""
+        L5_tested = L5.preprocess().spectralIndices(indices)
+        self.assertIsInstance(L5_tested, ee.image.Image)
+                
+    def test_L4_preprocess(self):
+        """Test the image module for Landsat 4"""
+        L4_tested = L4.preprocess().spectralIndices(indices)
+        self.assertIsInstance(L4_tested, ee.image.Image)
+    
+    # MODIS (TERRA + AQUA)
+    
+    def test_MCD43A4_preprocess(self):
+        """Test the image module for MCD43A4"""
+        MCD43A4_tested = MCD43A4.preprocess().spectralIndices(indices)
+        self.assertIsInstance(MCD43A4_tested, ee.image.Image)
+        
+    def test_MCD43A3_preprocess(self):
+        """Test the image module for MCD43A3"""
+        MCD43A3_tested = MCD43A3.preprocess()
+        self.assertIsInstance(MCD43A3_tested, ee.image.Image)
+        
+    def test_MCD43A1_preprocess(self):
+        """Test the image module for MCD43A1"""
+        MCD43A1_tested = MCD43A1.preprocess()
+        self.assertIsInstance(MCD43A1_tested, ee.image.Image)
+        
+    def test_MCD15A3H_preprocess(self):
+        """Test the image module for MCD15A3H"""
+        MCD15A3H_tested = MCD15A3H.preprocess()
+        self.assertIsInstance(MCD15A3H_tested, ee.image.Image)
+        
+    # MODIS (TERRA)
+        
+    def test_MOD09GQ_preprocess(self):
+        """Test the image module for MOD09GQ"""
+        MOD09GQ_tested = MOD09GQ.preprocess().spectralIndices(indices)
+        self.assertIsInstance(MOD09GQ_tested, ee.image.Image)
+        
+    def test_MOD10A1_preprocess(self):
[…]
