### 2021-05-01 davemlz/eemont#22 @giswqs
[JOSS Review] Comments and Suggestions
**Repo:**

- Suggest removing all .ipynb_checkpoints. You might want to add .ipynb_checkpoints to .gitignore so that these unnecessary folders won’t be pushed to GitHub. (see pull request https://github.com/davemlz/eemont/pull/21)
- Suggest adding the docs website (https://eemont.readthedocs.io) to the About section located at the upper-right corner of the repo page.
- Adding a Table of Content for README.rst, making it easier to navigate this long document.
- In README.rst, the `Contribution Steps` section probably belongs to a subsection of `Contributing`.
- You might want to consider publishing the package on conda-forge. Unless your package is available on conda-forge, other conda packages can’t list eermont as a dependency. Check out my tutorial on publishing a conda package if needed.
- The installation instructions can be improved by adding instructions such as how to update the package, install from GitHub, install from conda-forge (if it exists), etc.
- For all notebooks, `!pip install earthengine-api` is not needed because eemont/geemap already has earthengine-api as a dependency.
- Be cautious about using the GEE logo. Google is pretty picky about how their logos are used: https://www.google.com/permissions/logos-trademarks. I was warned before by a Google employee not to use the GEE logo. Unless you have received permission from Google to use/modify the GEE logo, you might want to design your own logo that is distinctive from the GEE logo.

**Paper:**

- The writing can be improved. I made some minor changes. See pull request https://github.com/davemlz/eemont/pull/21
