### 2021-05-01 davemlz/eemont#22 @giswqs
Related to https://github.com/openjournals/joss-reviews/issues/3168
