### 2021-05-01 #3168 @giswqs
@davemlz @kbarnhart I have provided my comments and suggestions at https://github.com/davemlz/eemont/issues/22 and [pull request](https://github.com/davemlz/eemont/pull/21/files#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424).
