### 2021-05-01 davemlz/eemont#19 @elbeejay
Great, I took a look at proposed changes in #21 and I think those are very good suggestions for changing some of the language and phrasing used.
