### 2021-05-02 davemlz/eemont#22 @davemlz
Hi, @giswqs!

Thank you very much for the suggestions and the PR! I'll merge it and start working on your comments and suggestions. I've planned to publish the package on conda-forge, but I have to figure out how to use a dependency that is not there, after that I'll publish it there!
