### 2021-05-02 davemlz/eemont#22 @giswqs
It seems all dependencies listed in requirements.txt are available on conda-forge. Which specific dependency are you referring to that's not available on conda-forge?
