### 2021-05-02 davemlz/eemont#22 @davemlz
You're right! It was my bad, I thought `geopy` wasn't on conda-forge.
