### 2021-05-02 447d823
README and Docs updated (#22)
