### 2021-05-03 davemlz/eemont#22 @davemlz
Hi, @giswqs!

I have updated the README (and also the docs and tutorials) following your suggestions (https://github.com/davemlz/eemont/commit/447d8237ce144e769476b2dbf7c97cb7b8f8b549, https://github.com/davemlz/eemont/commit/4513fb714c86d090d2de8b4818d7306c3cec286f).

I have also submitted the package to conda-forge (All checks passed) and I'm waiting for the review!

Thank you very much for your suggestions, comments and contributions!
