+++ eemont/common.py
@@ -5,6 +5,8 @@ import os
 import warnings
 import requests
 from box import Box
+from geopy.geocoders import get_geocoder_for_service
+import re
 
 warnings.simplefilter("always", UserWarning)
 
@@ -38,13 +40,9 @@ def _get_platform_STAC(args):
 
     for platform in platforms:
 
-        if eeDict[platform]["gee:type"] == "image_collection" and isinstance(
-            args, ee.image.Image
-        ):
+        if eeDict[platform]["gee:type"] == "image_collection" and isinstance(args, ee.image.Image):
             pltID = "/".join(ID.split("/")[:-1])
-        elif eeDict[platform]["gee:type"] == "image" and isinstance(
-            args, ee.imagecollection.ImageCollection
-        ):
+        elif eeDict[platform]["gee:type"] == "image" and isinstance(args, ee.imagecollection.ImageCollection):
             pass
         else:
             pltID = ID
@@ -195,9 +193,7 @@ def _get_expression_map(img, platformDict):
     }
 
     if platformDict["platform"] not in list(lookupPlatform.keys()):
-        raise Exception(
-            "Sorry, satellite platform not supported for index computation!"
-        )
+        raise Exception("Sorry, satellite platform not supported for index computation!")
 
     return lookupPlatform[platformDict["platform"]](img)
 
@@ -220,9 +216,7 @@ def _get_indices(online):
             "https://raw.githubusercontent.com/davemlz/awesome-ee-spectral-indices/main/output/spectral-indices-dict.json"
         ).json()
     else:
-        eemontDir = os.path.dirname(
-            pkg_resources.resource_filename("eemont", "eemont.py")
-        )
+        eemontDir = os.path.dirname(pkg_resources.resource_filename("eemont", "eemont.py"))
         dataPath = os.path.join(eemontDir, "data/spectral-indices-dict.json")
         f = open(dataPath)
         indices = json.load(f)
@@ -388,9 +382,7 @@ def _index(self, index, G, C1, C2, L, kernel, sigma, p, c, online):
 
     for idx in index:
         if idx not in list(spectralIndices.keys()):
-            warnings.warn(
-                "Index " + idx + " is not a built-in index and it won't be computed!"
-            )
+            warnings.warn("Index " + idx + " is not a built-in index and it won't be computed!")
         else:
 
             def temporalIndex(img):
@@ -399,21 +391,10 @@ def _index(self, index, G, C1, C2, L, kernel, sigma, p, c, online):
                 kernelParameters = _get_kernel_parameters(img, lookupDic, kernel, sigma)
                 lookupDic = {**lookupDic, **kernelParameters}
                 lookupDicCurated = _remove_none_dict(lookupDic)
-                if all(
-                    band in list(lookupDicCurated.keys())
-                    for band in spectralIndices[idx]["bands"]
-                ):
-                    return img.addBands(
-                        img.expression(
-                            spectralIndices[idx]["formula"], lookupDicCurated
-                        ).rename(idx)
-                    )
+                if all(band in list(lookupDicCurated.keys()) for band in spectralIndices[idx]["bands"]):
+                    return img.addBands(img.expression(spectralIndices[idx]["formula"], lookupDicCurated).rename(idx))
                 else:
-                    warnings.warn(
-                        "This platform doesn't have the required bands for "
-                        + idx
-                        + " computation!"
-                    )
+                    warnings.warn("This platform doesn't have the required bands for " + idx + " computation!")
                     return img
 
[…]
