### 2021-05-04 08f98ad
Implement ee.Geometry.Point from plus code and create placeholder methods for other geometries
