+++ eemont/common.py
@@ -16,12 +16,10 @@ warnings.simplefilter("always", UserWarning)
 
 def _get_platform_STAC(args):
     """Gets the platform (satellite) of an image (or image collection) and wheter if it is a Surface Reflectance product.
-
     Parameters
     ----------
     args : ee.Image | ee.ImageCollection
         Image or image collection to get the platform from.
-
     Returns
     -------
     dict
@@ -40,9 +38,13 @@ def _get_platform_STAC(args):
 
     for platform in platforms:
 
-        if eeDict[platform]["gee:type"] == "image_collection" and isinstance(args, ee.image.Image):
+        if eeDict[platform]["gee:type"] == "image_collection" and isinstance(
+            args, ee.image.Image
+        ):
             pltID = "/".join(ID.split("/")[:-1])
-        elif eeDict[platform]["gee:type"] == "image" and isinstance(args, ee.imagecollection.ImageCollection):
+        elif eeDict[platform]["gee:type"] == "image" and isinstance(
+            args, ee.imagecollection.ImageCollection
+        ):
             pass
         else:
             pltID = ID
@@ -63,12 +65,10 @@ def _get_platform_STAC(args):
 
 def _getSTAC(args):
     """Gets the STAC of the specified platform.
-
     Parameters
     ----------
     args : ee.Image | ee.ImageCollection
         Image or image collection to get the STAC from.
-
     Returns
     -------
     dict
@@ -93,14 +93,12 @@ def _getSTAC(args):
 
 def _get_expression_map(img, platformDict):
     """Gets the dictionary required for the map parameter in ee.Image.expression() method.
-
     Parameters
     ----------
     img : ee.Image
         Image to get the dictionary from.
     platformDict : dict
         Dictionary retrieved from the _get_platform() method.
-
     Returns
     -------
     dict
@@ -193,19 +191,19 @@ def _get_expression_map(img, platformDict):
     }
 
     if platformDict["platform"] not in list(lookupPlatform.keys()):
-        raise Exception("Sorry, satellite platform not supported for index computation!")
+        raise Exception(
+            "Sorry, satellite platform not supported for index computation!"
+        )
 
     return lookupPlatform[platformDict["platform"]](img)
 
 
 def _get_indices(online):
     """Retrieves the dictionary of indices used for the index() method in ee.Image and ee.ImageCollection classes.
-
     Parameters
     ----------
     online : boolean
         Wheter to retrieve the most recent list of indices directly from the GitHub repository and not from the local copy.
-
     Returns
     -------
     dict
@@ -216,7 +214,9 @@ def _get_indices(online):
             "https://raw.githubusercontent.com/davemlz/awesome-ee-spectral-indices/main/output/spectral-indices-dict.json"
         ).json()
     else:
-        eemontDir = os.path.dirname(pkg_resources.resource_filename("eemont", "eemont.py"))
+        eemontDir = os.path.dirname(
+            pkg_resources.resource_filename("eemont", "eemont.py")
+        )
         dataPath = os.path.join(eemontDir, "data/spectral-indices-dict.json")
         f = open(dataPath)
         indices = json.load(f)
@@ -226,7 +226,6 @@ def _get_indices(online):
 
 def _get_kernel_image(img, lookup, kernel, sigma, a, b):
     """Creates an ee.Image representing a kernel computed on bands [a] and [b].
-
     Parameters
     ----------
     img : ee.Image
@@ -241,7 +240,6 @@ def _get_kernel_image(img, lookup, kernel, sigma, a, b):
         Key of the first band to use.
     b : str
         Key of the second band to use.
-
     Returns
     -------
     ee.Image
@@ -265,11 +263,9 @@ def _get_kernel_image(img, lookup, kernel, sigma, a, b):
 
[…]
