+++ eemont/geometry.py
@@ -17,9 +17,11 @@ def _extend_eeGeometry():
 @_extend_staticmethod_eeGeometry()
 def BBoxFromQuery(query, geocoder="nominatim", **kwargs):
     """Constructs an ee.Geometry describing a bounding box from a query submitted to a geodocer using the geopy package.
+
     Tip
     ----------
     Check more info about constructors in the :ref:`User Guide<Constructors>`.
+
     Parameters
     ----------
     query : str
@@ -29,14 +31,17 @@ def BBoxFromQuery(query, geocoder="nominatim", **kwargs):
     **kwargs :
         Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
         user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
+
     Returns
     -------
     ee.Geometry.Polygon
         Geometry describing a bounding box from the specified query.
+
     See Also
     --------
     PointFromQuery : Constructs an ee.Geometry describing a point from a query submitted to a geodocer using the geopy package.
     MultiPointFromQuery : Constructs an ee.Geometry describing a multi-point from a query submitted to a geodocer using the geopy package.
+
     Examples
     --------
     >>> import ee, eemont
@@ -102,9 +107,11 @@ def BBoxFromQuery(query, geocoder="nominatim", **kwargs):
 @_extend_staticmethod_eeGeometry()
 def PointFromQuery(query, geocoder="nominatim", **kwargs):
     """Constructs an ee.Geometry describing a point from a query submitted to a geodocer using the geopy package. This returns exactly one pair of coordinates.
+
     Tip
     ----------
     Check more info about constructors in the :ref:`User Guide<Constructors>`.
+
     Parameters
     ----------
     query : str
@@ -114,14 +121,17 @@ def PointFromQuery(query, geocoder="nominatim", **kwargs):
     **kwargs :
         Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
         user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
+
     Returns
     -------
     ee.Geometry.Point
         Geometry describing a point from the specified query.
+
     See Also
     --------
     BBoxFromQuery : Constructs an ee.Geometry describing a bounding box from a query submitted to a geodocer using the geopy package.
     MultiPointFromQuery : Constructs an ee.Geometry describing a multi-point from a query submitted to a geodocer using the geopy package.
+
     Examples
     --------
     >>> import ee, eemont
@@ -154,9 +164,11 @@ def PointFromQuery(query, geocoder="nominatim", **kwargs):
 @_extend_staticmethod_eeGeometry()
 def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
     """Constructs an ee.Geometry describing a multi-point from a query submitted to a geodocer using the geopy package. This returns all pairs of coordinates retrieved by the query.
+
     Tip
     ----------
     Check more info about constructors in the :ref:`User Guide<Constructors>`.
+
     Parameters
     ----------
     query : str
@@ -166,14 +178,17 @@ def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
     **kwargs :
         Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
         user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
+
     Returns
     -------
     ee.Geometry.MultiPoint
         Geometry describing a multi-point from the specified query.
+
     See Also
     --------
[…]
