### 2021-05-04 074e9ae
Add tests for PointFromPlusCode (#23)
