### 2021-05-05 adc2996
Add function for converting llnglat to plus code and split openlocationcode import into separate function (#23)
