### 2021-05-05 60837b5
Set default codelength and implement Geometry to plus code conversion with tests (#23)
