### 2021-05-05 d02e6cf
Implement Feature to plus code conversion with tests (#23)
