+++ eemont/common.py
@@ -983,14 +983,10 @@ def _load_openlocationcode():
     """Attempt to load the openlocationcode.openlocationcode module and return it. Because the package is not available 
     through conda-forge, it cannot be made an installation dependency of eemont, so it is only loaded if needed. 
 
-    Parameters
-    ----------
-
     Returns
     -------
     module
         The openlocationcode.openlocationcode module.
-
     """
     try:
         from openlocationcode import openlocationcode
@@ -1009,11 +1005,11 @@ def _convert_lnglat_to_pluscode(lng, lat, codeLength):
         Latitude.  
     codeLength : int
         The number of significant digits in the output code, between 2 and 15. Shorter codes are less precise.
+    
     Returns
     -------
     str
         The plus code represented by the coordinate
-
     """
     olc = _load_openlocationcode()
 
@@ -1031,7 +1027,6 @@ def _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs):
     -------
     tuple
         The longitude and latitude of the plus code centroid.
-
     """
     olc = _load_openlocationcode()
 
