### 2021-05-05 4a54cfc
Implement and test MultiPointFromPlusCode (#23)
