### 2021-05-05 2428f37
conda-forge added (#22)
