### 2021-05-05 davemlz/eemont#22 @davemlz
Hi, @giswqs!

`eemont` is now on `conda-forge`! I have updated the README and the Docs with the installation guidelines.
