+++ docs/index.rst
@@ -15,15 +15,12 @@ Welcome to eemont!
         :alt: Documentation Status
 
 .. image:: https://github.com/davemlz/eemont/actions/workflows/tests.yml/badge.svg
-        :target: https://github.com/davemlz/eemont/actions/workflows/tests.yml        
-
-.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
-    :target: https://github.com/psf/black
-
+        :target: https://github.com/davemlz/eemont/actions/workflows/tests.yml  
+        
 .. image:: https://img.shields.io/badge/Donate-buy%20me%20a%20coffee-yellow.svg
         :target: https://www.buymeacoffee.com/davemlz
         
-.. image:: https://static.pepy.tech/personalized-badge/eemont?period=total&units=international_system&left_color=grey&right_color=lightgrey&left_text=Downloads
+.. image:: https://static.pepy.tech/personalized-badge/eemont?period=total&units=international_system&left_color=grey&right_color=green&left_text=Downloads
         :target: https://pepy.tech/project/eemont
         
 .. image:: https://img.shields.io/badge/GEE%20Community-Developer%20Resources-00b6ff.svg
@@ -31,6 +28,12 @@ Welcome to eemont!
         
 .. image:: https://img.shields.io/twitter/follow/dmlmont?style=social
         :target: https://twitter.com/dmlmont
+        
+.. image:: https://joss.theoj.org/papers/34696c5b62c50898b4129cbbe8befb0a/status.svg
+    :target: https://joss.theoj.org/papers/34696c5b62c50898b4129cbbe8befb0a
+        
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
 
 **A Python package that extends** `Google Earth Engine <https://earthengine.google.com/>`_
 
