### 2021-05-08 9ae048e
Implement other geometry constructors (#23)
