### 2021-05-08 0b78cf8
Make polygon feature test match geometry test and fix docstrings
