+++ eemont/geometry.py
@@ -244,78 +244,109 @@ def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
 
 @_extend_staticmethod_eeGeometry()
 def MultiPointFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in pluscodes]
+    coordinates = [
+        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+        for pluscode in pluscodes
+    ]
     return ee.Geometry.MultiPoint(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def PolygonFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in pluscodes]
+    coordinates = [
+        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+        for pluscode in pluscodes
+    ]
     return ee.Geometry.Polygon(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def MultiPolygonFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [[_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in code_group] for code_group in pluscodes]
+    coordinates = [
+        [
+            _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+            for pluscode in code_group
+        ]
+        for code_group in pluscodes
+    ]
     return ee.Geometry.MultiPolygon(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def LineStringFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in pluscodes]
+    coordinates = [
+        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+        for pluscode in pluscodes
+    ]
     return ee.Geometry.LineString(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def MultiLineStringFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [[_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in code_group] for code_group in pluscodes]
+    coordinates = [
+        [
+            _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+            for pluscode in code_group
+        ]
+        for code_group in pluscodes
+    ]
     return ee.Geometry.MultiLineString(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def LinearRingFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in pluscodes]
+    coordinates = [
+        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+        for pluscode in pluscodes
+    ]
     return ee.Geometry.LinearRing(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
 def RectangleFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [_convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs) for pluscode in pluscodes]
+    coordinates = [
[…]
