### 2021-05-08 c1cdd8b
Reimplement geometry to plus code conversion to handle nested coordinates
