### 2021-05-08 98b951d
Add test coverage for converting all geometries to plus codes
