+++ tests/test_geometry.py
@@ -103,13 +103,24 @@ class Test(unittest.TestCase):
         test = ee.Geometry.RectangleFromPlusCode(codes)
         self.assertIsInstance(test, ee.geometry.Geometry)
 
-    def test_pluscode_from_point(self):
+    def test_pluscode_from_Point(self):
         """Test that an ee.Geometry.Point returns a plus code string"""
-        pt = ee.Geometry.Point([-105, 40])
-        test = pt.plusCode()
+        geom = ee.Geometry.Point([-105, 40])
+        test = geom.plusCode()
         self.assertIsInstance(test, str)
+    
+    def test_pluscode_from_MultiPoint(self):
+        """Test that an ee.Geometry.MultiPoint returns a plus code list of correct length"""
+        coords = [
+            [-105, 40],
+            [-105, 39],
+            [-104, 39],
+        ]
+        geom = ee.Geometry.MultiPoint(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), len(coords))
 
-    def test_pluscode_from_polygon(self):
+    def test_pluscode_from_Polygon(self):
         """Test that an ee.Geometry.Polygon returns a plus code list of correct length"""
         coords = [
             [-105, 40],
@@ -117,9 +128,62 @@ class Test(unittest.TestCase):
             [-104, 39],
             [-105, 40],
         ]
-        poly = ee.Geometry.Polygon(coords)
-        test = poly.plusCode()
+        geom = ee.Geometry.Polygon(coords)
+        test = geom.plusCode()
         self.assertEqual(len(test), len(coords))
+
+    def test_pluscode_from_MultiPolygon(self):
+        """Test that an ee.Geometry.MultiPolygon returns a plus code list of correct length"""
+        coords = [
+            [[-105, 40], [-105, 39], [-104, 39]],
+            [[-104, 41], [-104, 40], [-103, 40], [-103, 39]]
+        ]
+        geom = ee.Geometry.MultiPolygon(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), len(coords))
+
+    def test_pluscode_from_LineString(self):
+        """Test that an ee.Geometry.LineString returns a plus code list of correct length"""
+        coords = [
+            [-105, 40],
+            [-105, 39],
+            [-104, 39],
+        ]
+        geom = ee.Geometry.LineString(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), len(coords))
+
+    def test_pluscode_from_MultiLineString(self):
+        """Test that an ee.Geometry.MultiLineString returns a plus code list of correct length"""
+        coords = [
+            [[-105, 40], [-105, 39], [-104, 39]],
+            [[-104, 41], [-104, 40], [-103, 40], [-103, 39]]
+        ]
+        geom = ee.Geometry.MultiLineString(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), len(coords))
+
+    def test_pluscode_from_LinearRing(self):
+        """Test that an ee.Geometry.LinearRing returns a plus code list of correct length"""
+        coords = [
+            [-105, 40],
+            [-105, 39],
+            [-104, 39],
+            [-105, 40]
+        ]
+        geom = ee.Geometry.LinearRing(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), len(coords))
+
+    def test_pluscode_from_Rectangle(self):
+        """Test that an ee.Geometry.Rectangle returns a plus code list of correct length"""
+        coords = [
+            [-105, 40],
+            [-104, 39]
+        ]
+        geom = ee.Geometry.Rectangle(coords)
+        test = geom.plusCode()
+        self.assertEqual(len(test), 5)
 
 
 if __name__ == '__main__':
