### 2021-05-09 827f8f6
Refactor geometry constructors to implement the common plus code conversion function that handles nested coordinates. Make pluscodes plural where relevant. Modify the pluscode tests to ensure that pluscodes returned by ee.Geometry.plusCodes and ee.Feature.plusCodes can be used as geometry constructors. (#23)
