+++ eemont/geometry.py
@@ -1,7 +1,7 @@
 import ee
 import geopy
 from geopy.geocoders import get_geocoder_for_service
-from eemont.common import _convert_pluscode_to_lnglat, _convert_lnglat_to_pluscode, _convert_nested_lnglat_to_pluscode
+from eemont.common import _convert_lnglats_to_pluscodes, _convert_pluscodes_to_lnglats
 
 
 def _extend_staticmethod_eeGeometry():
@@ -237,81 +237,54 @@ def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
 
 @_extend_staticmethod_eeGeometry()
 def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
-    coordinates = _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
+    coordinates = _convert_pluscodes_to_lnglats(pluscode, geocoder, **kwargs)
     return ee.Geometry.Point(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
-def MultiPointFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [
-        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
-        for pluscode in pluscodes
-    ]
+def MultiPointFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
+    coordinates = _convert_pluscodes_to_lnglats(pluscodes, geocoder, **kwargs)
     return ee.Geometry.MultiPoint(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
-def PolygonFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [
-        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
-        for pluscode in pluscodes
-    ]
+def PolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
+    coordinates = _convert_pluscodes_to_lnglats(pluscodes, geocoder, **kwargs)
     return ee.Geometry.Polygon(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
-def MultiPolygonFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [
-        [
-            _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
-            for pluscode in code_group
-        ]
-        for code_group in pluscodes
-    ]
+def MultiPolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
+    coordinates = _convert_pluscodes_to_lnglats(pluscodes, geocoder, **kwargs)
     return ee.Geometry.MultiPolygon(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
-def LineStringFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [
-        _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
-        for pluscode in pluscodes
-    ]
+def LineStringFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
+    coordinates = _convert_pluscodes_to_lnglats(pluscodes, geocoder, **kwargs)
     return ee.Geometry.LineString(coordinates)
 
 
 @_extend_staticmethod_eeGeometry()
-def MultiLineStringFromPlusCode(pluscodes, geocoder="nominatim", **kwargs):
-    coordinates = [
-        [
-            _convert_pluscode_to_lnglat(pluscode, geocoder, **kwargs)
-            for pluscode in code_group
-        ]
-        for code_group in pluscodes
-    ]
[…]
