+++ tests/test_geometry.py
@@ -34,94 +34,94 @@ class Test(unittest.TestCase):
         test = ee.Geometry.PointFromPlusCode(code, user_agent = 'eemont-feature-test')
         self.assertIsInstance(test, ee.geometry.Geometry)
     
-    def test_MultiPointFromPlusCode(self):
-        """Test the MultiPointFromPlusCode constructor with a list of plus codes"""
+    def test_MultiPointFromPlusCodes(self):
+        """Test the MultiPointFromPlusCodes constructor with a list of plus codes"""
         codes = [
             '85FQ2222+22', 
             '85FR2222+22', 
             '85GR2222+22', 
         ]
-        test = ee.Geometry.MultiPointFromPlusCode(codes)
+        test = ee.Geometry.MultiPointFromPlusCodes(codes)
         self.assertIsInstance(test, ee.geometry.Geometry)
     
-    def test_PolygonFromPlusCode(self):
-        """Test the PolygonFromPlusCode constructor with a list of plus codes"""
+    def test_PolygonFromPlusCodes(self):
+        """Test the PolygonFromPlusCodes constructor with a list of plus codes"""
         codes = [
             '85FQ2222+22', 
             '85FR2222+22', 
             '85GR2222+22', 
         ]
-        test = ee.Geometry.PolygonFromPlusCode(codes)
+        test = ee.Geometry.PolygonFromPlusCodes(codes)
         self.assertIsInstance(test, ee.geometry.Geometry)
 
-    def test_MultiPolygonFromPlusCode(self):
-        """Test the MultiPolygonFromPlusCode constructor with a list of lists of plus codes"""
+    def test_MultiPolygonFromPlusCodes(self):
+        """Test the MultiPolygonFromPlusCodes constructor with a list of lists of plus codes"""
         codes = [
             ['85FQ2222+22', '85FR2222+22', '85GR2222+22'], 
             ['85FP8PC2+G2', '85FPJF23+G4', '85FPMW2R+RP'],
             ['85FPRJ5W+82', '85GP3M67+CP', '85GQ2R7C+38', '85FQC5X6+FJ', '85FPPXQJ+F3'],
         ]
-        test = ee.Geometry.MultiPolygonFromPlusCode(codes)
+        test = ee.Geometry.MultiPolygonFromPlusCodes(codes)
         self.assertIsInstance(test, ee.geometry.Geometry)
 
-    def test_LineStringFromPlusCode(self):
-        """Test the LineStringFromPlusCode constructor with a list of plus codes"""
+    def test_LineStringFromPlusCodes(self):
+        """Test the LineStringFromPlusCodes constructor with a list of plus codes"""
         codes = [
             '85FQ2222+22', 
             '85FR2222+22', 
             '85GR2222+22', 
         ]
-        test = ee.Geometry.LineStringFromPlusCode(codes)
+        test = ee.Geometry.LineStringFromPlusCodes(codes)
         self.assertIsInstance(test, ee.geometry.Geometry)
 
-    def test_MultiLineStringFromPlusCode(self):
-        """Test the MultiLineStringFromPlusCode constructor with a list of lists of plus codes"""
+    def test_MultiLineStringFromPlusCodes(self):
+        """Test the MultiLineStringFromPlusCodes constructor with a list of lists of plus codes"""
         codes = [
             ['85FQ2222+22', '85FR2222+22', '85GR2222+22'], 
             ['85FP8PC2+G2', '85FPJF23+G4', '85FPMW2R+RP'],
             ['85FPRJ5W+82', '85GP3M67+CP', '85GQ2R7C+38', '85FQC5X6+FJ', '85FPPXQJ+F3'],
         ]
[…]

[…]
