+++ eemont/imagecollection.py
@@ -378,7 +378,7 @@ def index(
 
     Tip
     ----------
-    Check more info about the supported platforms and spectral indices in the :ref:`User Guide<Spectral Indices>`.
+    Check more info about the supported platforms and spectral indices in the :ref:`User Guide<Spectral Indices Computation>`.
 
     Parameters
     ----------
@@ -525,7 +525,7 @@ def spectralIndices(
 
     Tip
     ----------
-    Check more info about the supported platforms and spectral indices in the :ref:`User Guide<Spectral Indices>`.
+    Check more info about the supported platforms and spectral indices in the :ref:`User Guide<Spectral Indices Computation>`.
 
     Parameters
     ----------
