### 2021-05-14 208ab66
cexp added to spectralIndices
