### 2021-05-14 9343be2
nexp added to spectralIndices
