### 2021-05-14 0ccf3a1
alpha added to spectralIndices
