+++ eemont/image.py
@@ -593,6 +593,7 @@ def index(
     L=1.0,
     cexp=1.16,
     nexp=2.0,
+    alpha=0.1,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
@@ -674,6 +675,8 @@ def index(
         Exponent used for OCVI.
     nexp : float, default = 2.0
         Exponent used for GDVI.
+    alpha : float, default = 0.1
+        Weighting coefficient  used for WDRVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -737,7 +740,7 @@ def index(
         PendingDeprecationWarning,
     )
 
-    return _index(self, index, G, C1, C2, L, cexp, nexp, kernel, sigma, p, c, online)
+    return _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, online)
 
 
 @_extend_eeImage()
@@ -750,6 +753,7 @@ def spectralIndices(
     L=1.0,
     cexp=1.16,
     nexp=2.0,
+    alpha=0.1,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
@@ -824,9 +828,11 @@ def spectralIndices(
     L : float, default = 1.0
         Canopy background adjustment. Used just for index = ['EVI','SAVI'].
     cexp : float, default = 1.16
-        Exponent used for OCVI.
+        Exponent used for OCVI.    
     nexp : float, default = 2.0
         Exponent used for GDVI.
+    alpha : float, default = 0.1
+        Weighting coefficient  used for WDRVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -883,7 +889,7 @@ def spectralIndices(
 
     >>> S2.spectralIndices('all')
     """
-    return _index(self, index, G, C1, C2, L, cexp, nexp, kernel, sigma, p, c, online)
+    return _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, online)
 
 
 @_extend_eeImage()
