+++ eemont/imagecollection.py
@@ -366,6 +366,7 @@ def index(
     L=1.0,
     cexp=1.16,
     nexp=2.0,
+    alpha=0.1,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
@@ -447,6 +448,8 @@ def index(
         Exponent used for OCVI.
     nexp : float, default = 2.0
         Exponent used for GDVI.
+    alpha : float, default = 0.1
+        Weighting coefficient  used for WDRVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -510,7 +513,7 @@ def index(
         PendingDeprecationWarning,
     )
 
-    return _index(self, index, G, C1, C2, L, cexp, nexp, kernel, sigma, p, c, online)
+    return _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, online)
 
 
 @_extend_eeImageCollection()
@@ -523,6 +526,7 @@ def spectralIndices(
     L=1.0,
     cexp=1.16,
     nexp=2.0,
+    alpha=0.1,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
@@ -600,6 +604,8 @@ def spectralIndices(
         Exponent used for OCVI.
     nexp : float, default = 2.0
         Exponent used for GDVI.
+    alpha : float, default = 0.1
+        Weighting coefficient  used for WDRVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -656,7 +662,7 @@ def spectralIndices(
 
     >>> S2.spectralIndices('all')
     """
-    return _index(self, index, G, C1, C2, L, cexp, nexp, kernel, sigma, p, c, online)
+    return _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, online)
 
 
 @_extend_eeImageCollection()
