### 2021-05-14 58a39c2
slope and intercept added to spectralIndices
