+++ eemont/common.py
@@ -327,7 +327,7 @@ def _get_kernel_parameters(img, lookup, kernel, sigma):
     return kernelParameters
 
 
-def _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, online):
+def _index(self, index, G, C1, C2, L, cexp, nexp, alpha, slope, intercept, kernel, sigma, p, c, online):
     """Computes one or more spectral indices (indices are added as bands) for an image oir image collection.
 
     Parameters
@@ -350,6 +350,10 @@ def _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, on
         Exponent used for GDVI.
     alpha : float
         Weighting coefficient  used for WDRVI.
+    slope : float
+        Soil line slope.
+    intercept : float
+        Soil line intercept.
     kernel : str
         Kernel used for kernel indices.
     sigma : str | float
@@ -383,6 +387,8 @@ def _index(self, index, G, C1, C2, L, cexp, nexp, alpha, kernel, sigma, p, c, on
         "cexp": float(cexp),
         "nexp": float(nexp),
         "alpha": float(alpha),
+        "slope": float(slope),
+        "intercept": float(intercept),
         "p": float(p),
         "c": float(c),
     }
