### 2021-05-15 435f302
slope and intercept modified
