### 2021-05-18 f502461
Fix geometry.plusCode bug by passing codeLength arg (#23)
