+++ eemont/feature.py
@@ -183,4 +183,4 @@ def plusCodes(self, codeLength=10):
     >>> pt.plusCodes()
     '85GQ2222+22'
     """
-    return self.geometry().plusCodes()
+    return self.geometry().plusCodes(codeLength)
