+++ eemont/image.py
@@ -10,14 +10,10 @@ from .common import _preprocess
 from .common import _getSTAC
 from .common import _getDOI
 from .common import _getCitation
+from .extending import extend
 
 
-def _extend_eeImage():
-    """Decorator. Extends the ee.ImageCollection class."""
-    return lambda f: (setattr(ee.image.Image, f.__name__, f) or f)
-
-
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __add__(self, other):
     """Computes the addition between two images.
 
@@ -36,7 +32,7 @@ def __add__(self, other):
     return self.add(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __radd__(self, other):
     """Computes the addition between two images.
 
@@ -55,7 +51,7 @@ def __radd__(self, other):
     return self.add(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __sub__(self, other):
     """Computes the subtraction between two images.
 
@@ -74,7 +70,7 @@ def __sub__(self, other):
     return self.subtract(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __rsub__(self, other):
     """Computes the subtraction between two images.
 
@@ -93,7 +89,7 @@ def __rsub__(self, other):
     return ee.Image(other).subtract(self)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __mul__(self, other):
     """Computes the multiplication between two images.
 
@@ -112,7 +108,7 @@ def __mul__(self, other):
     return self.multiply(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __rmul__(self, other):
     """Computes the multiplication between two images.
 
@@ -131,7 +127,7 @@ def __rmul__(self, other):
     return self.multiply(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __truediv__(self, other):
     """Computes the division between two images.
 
@@ -150,7 +146,7 @@ def __truediv__(self, other):
     return self.divide(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __rtruediv__(self, other):
     """Computes the division between two images.
 
@@ -169,7 +165,7 @@ def __rtruediv__(self, other):
     return ee.Image(other).divide(self)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __floordiv__(self, other):
     """Computes the floor division of two images.
 
@@ -188,7 +184,7 @@ def __floordiv__(self, other):
     return self.divide(other).floor()
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __rfloordiv__(self, other):
     """Computes the floor division of two images.
 
@@ -207,7 +203,7 @@ def __rfloordiv__(self, other):
     return ee.Image(other).divide(self).floor()
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __mod__(self, other):
     """Computes the modulo of two images.
 
@@ -226,7 +222,7 @@ def __mod__(self, other):
     return self.mod(other)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __rmod__(self, other):
     """Computes the modulo of two images.
 
@@ -245,7 +241,7 @@ def __rmod__(self, other):
     return ee.Image(other).mod(self)
 
 
-@_extend_eeImage()
+@extend(ee.image.Image)
 def __pow__(self, other):
     """Computes the base (left operand) to the power (right operand).
 
@@ -264,7 +260,7 @@ def __pow__(self, other):
[…]
