+++ eemont/featurecollection.py
@@ -2,24 +2,10 @@ import ee
 import geopy
 from geopy.geocoders import get_geocoder_for_service
 from .geometry import *
+from .extending import extend
 
 
-def _extend_staticmethod_eeFeatureCollection():
-    """Decorator. Extends the ee.FeatureCollection class with a static method."""
-    return lambda f: (
-        setattr(ee.featurecollection.FeatureCollection, f.__name__, staticmethod(f))
-        or f
-    )
-
-
-def _extend_eeFeatureCollection():
-    """Decorator. Extends the ee.FeatureCollection class."""
-    return lambda f: (
-        setattr(ee.featurecollection.FeatureCollection, f.__name__, f) or f
-    )
-
-
-@_extend_staticmethod_eeFeatureCollection()
+@extend(ee.featurecollection.FeatureCollection, static=True)
 def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
     """Constructs an ee.Feature describing a point from a query submitted to a geodocer using the geopy package. This returns all pairs of coordinates retrieved by the query.
     The properties of the feature collection correspond to the raw properties retrieved by the locations of the query.
