+++ eemont/common.py
@@ -5,6 +5,7 @@ import os
 import warnings
 import requests
 from box import Box
+from geopy.geocoders import get_geocoder_for_service
 
 warnings.simplefilter("always", UserWarning)
 
@@ -1019,3 +1020,52 @@ def _getCitation(args):
     eeDict = json.load(f)
 
     return eeDict[platformDict["platform"]]["sci:citation"]
+
+
+# Geocoding
+# --------------------------
+
+
+def _retrieve_location(query, geocoder, exactly_one, **kwargs):
+    """Retrieves a location from a query.
+
+    Parameters
+    ----------
+    query : str
+        Address, query or structured query to geocode.
+    geocoder : str
+        Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
+    exactly_one : boolean
+        Whether to retrieve just one location.
+    **kwargs :
+        Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
+        user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
+
+    Returns
+    -------
+    Location
+        Retrieved location.
+    """
+    cls = get_geocoder_for_service(geocoder)
+    geolocator = cls(**kwargs)
+    location = geolocator.geocode(query, exactly_one=exactly_one)
+    if location is None:
+        raise Exception("No matches were found for your query!")
+    else:
+        return location
+
+
+def _lnglat_from_location(location):
+    """Returns the longitude and latitude from a location.
+
+    Parameters
+    ----------
+    location : Location
+        Retrieved location. Must be only one location.
+
+    Returns
+    -------
+    tuple
+        The longitude and latitude geocoded from the query.
+    """
+    return [location.longitude, location.latitude]
