### 2021-05-22 2bd244e
Docstrings (#23)
