+++ eemont/geometry.py
@@ -215,55 +215,418 @@ def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
 
 @extend(ee.geometry.Geometry, static=True)
 def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
+    """Constructs an ee.Geometry describing a point from a Plus Code. 
+    
+    If the Plus Code is full, it will be decoded directly using the openlocationcode package. If it is a short Code with
+    a reference location, the reference will be geocoded using the geopy package.
+
+    Tip
+    ----------
+    Check more info about constructors in the :ref:`User Guide<Constructors>`.
+
+    Parameters
+    ----------
+    pluscode : str
+        A full Plus Code or a short Plus Code with a geocodable reference location appended to it. 
+    geocoder : str, default = 'nominatim'
+        Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
+    **kwargs :
+        Keywords arguments for geolocator.geocode(). The user_agent argument is mandatory (this argument can be set as user_agent = 'my-gee-username' or
+        user_agent = 'my-gee-app-name'). Please visit https://geopy.readthedocs.io/ for more info.
+
+    Returns
+    -------
+    ee.Geometry.Point
+        Geometry describing a point from the Plus Code centroid.
+
+    See Also
+    --------
+    MultiPointFromPlusCodes : Constructs an ee.Geometry describing multiple points from a list of Plus Codes.
+    PolygonFromPlusCodes : Constructs an ee.Geometry describing a polygon from a list of Plus Codes.
+    MultiPolygonFromPlusCodes : Constructs an ee.Geometry describing multiple polygons from a list of lists of Plus Codes.
+    LineStringFromPlusCodes : Constructs an ee.Geometry describing a line from a list of Plus Codes.
+    MultiLineStringFromPlusCodes : Constructs an ee.Geometry describing multiple lines from a list of lists of Plus Codes.
+    LinearRingFromPlusCodes : Constructs an ee.Geometry describing a linear ring from a list of Plus Codes.
+    RectangleFromPlusCodes : Constructs an ee.Geometry describing a rectangle from a list of two Plus Code corners.
+
+    Examples
+    --------
+    >>> import ee, eemont
+    >>> ee.Authenticate()
+    >>> ee.Initialize()
+    >>> geom = ee.Geometry.PointFromPlusCode('QXGV+XH Denver, CO, USA',user_agent = 'my-gee-eemont-query')
+    >>> geom.coordinates().getInfo()
+    [-105.0060625, 39.777437500000005]
+    """
     coordinates = _convert_pluscodes_to_lnglats(pluscode, geocoder, **kwargs)
     return ee.Geometry.Point(coordinates)
 
 
 @extend(ee.geometry.Geometry, static=True)
 def MultiPointFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
+    """Constructs an ee.Geometry describing multiple points from a list of Plus Codes. 
+    
+    If the Plus Codes are full, they will be decoded directly using the openlocationcode package. If they are short Codes with
+    reference locations, the references will be geocoded using the geopy package.
+
+    Tip
+    ----------
+    Check more info about constructors in the :ref:`User Guide<Constructors>`.
+
+    Parameters
+    ----------
+    pluscodes : str
+        A list of full Plus Codes or short Plus Codes with geocodable reference locations appended to them. 
+    geocoder : str, default = 'nominatim'
+        Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
+    **kwargs :
[…]
