### 2021-05-22 f6b7ab0
Test plusCode method of all features and update method to avoid deprecation warning (#23)
