+++ tests/test_feature.py
@@ -17,8 +17,27 @@ class Test(unittest.TestCase):
         test = ee.Feature.PointFromQuery('Colombia',user_agent = 'eemon-feature-test')
         self.assertIsInstance(test, ee.feature.Feature)
 
-    def test_pluscodes_from_Polygon(self):
-        """Test that the Plus Codes returned by a Polygon feature can construct a new Polygon geometry"""
+    def test_pluscodes_from_Point_feature(self):
+        """Test that the plus codes returned by a Point feature can construct a new Point"""
+        geom = ee.Geometry.Point([-105, 40])
+        feature = ee.Feature(geom)
+        test = ee.Geometry.PointFromPlusCode(feature.plusCodes())
+        self.assertIsInstance(test, ee.geometry.Geometry)
+    
+    def test_pluscodes_from_MultiPoint_feature(self):
+        """Test that the plus codes returned by a MultiPoint feature can construct a new MultiPoint"""
+        coords = [
+            [-105, 40],
+            [-105, 39],
+            [-104, 39],
+        ]
+        geom = ee.Geometry.MultiPoint(coords)
+        feature = ee.Feature(geom)
+        test = ee.Geometry.MultiPointFromPlusCodes(feature.plusCodes())
+        self.assertIsInstance(test, ee.geometry.Geometry)
+
+    def test_pluscodes_from_Polygon_feature(self):
+        """Test that the plus codes returned by a Polygon feature can construct a new Polygon"""
         coords = [
             [-105, 40],
             [-105, 39],
@@ -29,6 +48,68 @@ class Test(unittest.TestCase):
         feature = ee.Feature(geom)
         test = ee.Geometry.PolygonFromPlusCodes(feature.plusCodes())
         self.assertIsInstance(test, ee.geometry.Geometry)
+
+    def test_pluscodes_from_MultiPolygon_feature(self):
+        """Test that the plus codes returned by a MultiPolygon feature can construct a new MultiPolygon"""
+        coords = [
+            [[-105, 40], [-105, 39], [-104, 39]],
+            [[-104, 41], [-104, 40], [-103, 40], [-103, 39]]
+        ]
+        geom = ee.Geometry.MultiPolygon(coords)
+        feature = ee.Feature(geom)
+        test = ee.Geometry.MultiPolygonFromPlusCodes(feature.plusCodes())
+        self.assertIsInstance(test, ee.geometry.Geometry)
+
+    def test_pluscodes_from_LineString_feature(self):
+        """Test that the plus codes returned by a LineString feature can construct a new LineString"""        
+        coords = [
+            [-105, 40],
+            [-105, 39],
+            [-104, 39],
+        ]
+        geom = ee.Geometry.LineString(coords)
+        feature = ee.Feature(geom)
+        test = ee.Geometry.LineStringFromPlusCodes(feature.plusCodes())
+        self.assertIsInstance(test, ee.geometry.Geometry)
+
+    def test_pluscodes_from_MultiLineString_feature(self):
+        """Test that the plus codes returned by a MultiLineString feature can construct a new MultiLineString"""
+        coords = [
+            [[-105, 40], [-105, 39], [-104, 39]],
+            [[-104, 41], [-104, 40], [-103, 40], [-103, 39]]
+        ]
+        geom = ee.Geometry.MultiLineString(coords)
+        feature = ee.Feature(geom)
+        test = ee.Geometry.MultiLineStringFromPlusCodes(feature.plusCodes())
+        self.assertIsInstance(test, ee.geometry.Geometry)
+
+    def test_pluscodes_from_LinearRing_feature(self):
+        """Test that the plus codes returned by a LinearRing feature can construct a new LinearRing"""
+        coords = [
+            [-105, 40],
[…]
