+++ eemont/geometry.py
@@ -215,8 +215,8 @@ def MultiPointFromQuery(query, geocoder="nominatim", **kwargs):
 
 @extend(ee.geometry.Geometry, static=True)
 def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
-    """Constructs an ee.Geometry describing a point from a Plus Code. 
-    
+    """Constructs an ee.Geometry describing a point from a Plus Code.
+
     If the Plus Code is full, it will be decoded directly using the openlocationcode package. If it is a short Code with
     a reference location, the reference will be geocoded using the geopy package.
 
@@ -227,7 +227,7 @@ def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
     Parameters
     ----------
     pluscode : str
-        A full Plus Code or a short Plus Code with a geocodable reference location appended to it. 
+        A full Plus Code or a short Plus Code with a geocodable reference location appended to it.
     geocoder : str, default = 'nominatim'
         Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
     **kwargs :
@@ -264,8 +264,8 @@ def PointFromPlusCode(pluscode, geocoder="nominatim", **kwargs):
 
 @extend(ee.geometry.Geometry, static=True)
 def MultiPointFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
-    """Constructs an ee.Geometry describing multiple points from a list of Plus Codes. 
-    
+    """Constructs an ee.Geometry describing multiple points from a list of Plus Codes.
+
     If the Plus Codes are full, they will be decoded directly using the openlocationcode package. If they are short Codes with
     reference locations, the references will be geocoded using the geopy package.
 
@@ -276,7 +276,7 @@ def MultiPointFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
     Parameters
     ----------
     pluscodes : str
-        A list of full Plus Codes or short Plus Codes with geocodable reference locations appended to them. 
+        A list of full Plus Codes or short Plus Codes with geocodable reference locations appended to them.
     geocoder : str, default = 'nominatim'
         Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
     **kwargs :
@@ -314,8 +314,8 @@ def MultiPointFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
 
 @extend(ee.geometry.Geometry, static=True)
 def PolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
-    """Constructs an ee.Geometry describing a polygon from a list of Plus Codes. 
-    
+    """Constructs an ee.Geometry describing a polygon from a list of Plus Codes.
+
     If the Plus Codes are full, they will be decoded directly using the openlocationcode package. If they are short Codes with
     reference locations, the references will be geocoded using the geopy package.
 
@@ -326,7 +326,7 @@ def PolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
     Parameters
     ----------
     pluscodes : str
-        A list of full Plus Codes or short Plus Codes with geocodable reference locations appended to them. 
+        A list of full Plus Codes or short Plus Codes with geocodable reference locations appended to them.
     geocoder : str, default = 'nominatim'
         Geocoder to use. Please visit https://geopy.readthedocs.io/ for more info.
     **kwargs :
@@ -366,8 +366,8 @@ def PolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
 
 @extend(ee.geometry.Geometry, static=True)
 def MultiPolygonFromPlusCodes(pluscodes, geocoder="nominatim", **kwargs):
-    """Constructs an ee.Geometry describing multiple polygons from a list of lists of Plus Codes. 
-    
[…]
