### 2021-05-23 2f27138
Fix docstring (#23)
