### 2021-05-24 #3168 @patrickcgray
My apologies for my review taking the full six weeks. I've gone through the paper and the repo and have finished my review. I recommend an acceptance and have a few minor comments raised in this [issue](https://github.com/davemlz/eemont/issues/28). I was not able to log in as a reviewer for some reason, but have included my checklist below. @kbarnhart if you're able to resubmit the invite I can log in and take care of fit in the proper format if you'd like. Otherwise everything looks great to me and I think this is a excellent addition to the community.

- [x]  I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
- [x]  I confirm that I read and will adhere to the JOSS code of conduct.
- [x]  Repository: Is the source code for this software available at the repository url?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [x] Contribution and authorship: Has the submitting author (@davemlz) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines
- [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
- [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
