+++ docs/classes/stubs/eemont.image.getDOI.rst
@@ -1,0 +1,6 @@
+eemont.image.getDOI
+===================
+
+.. currentmodule:: eemont.image
+
+.. autofunction:: getDOI
