+++ docs/classes/stubs/eemont.imagecollection.getSTAC.rst
@@ -1,0 +1,6 @@
+eemont.imagecollection.getSTAC
+==============================
+
+.. currentmodule:: eemont.imagecollection
+
+.. autofunction:: getSTAC
