+++ README.rst
@@ -37,6 +37,12 @@ eemont
     
 **A Python package that extends** `Google Earth Engine <https://earthengine.google.com/>`_
 
+- GitHub: `https://github.com/davemlz/eemont <https://github.com/davemlz/eemont>`_
+- Documentation: `https://eemont.readthedocs.io/ <https://eemont.readthedocs.io/>`_
+- PyPI: `https://pypi.org/project/eemont/ <https://pypi.org/project/eemont/>`_
+- Conda-Forge: `https://anaconda.org/conda-forge/eemont <https://anaconda.org/conda-forge/eemont>`_
+- Tutorials: `https://github.com/davemlz/eemont/tree/master/tutorials <https://github.com/davemlz/eemont/tree/master/tutorials>`_
+
 **Table of Contents**
 
 - `Overview`_
