### 2021-05-26 davemlz/eemont#28 @davemlz
Hi, @patrickcgray!

I have updated the paper following your suggested template for the Statement of Need and added some links to the README (and also to the main page of the Docs), including a link to the tutorials! :)

Thank you for the comments and suggestions again, @patrickcgray!
