### 2021-05-26 #3168 @kbarnhart
@davemlz thank you for addressing the reviewer comments.

@patrickcgray thank you for indicating that you think the submission is ready to be accepted.

@elbeejay @giswqs please revisit this review now that the author has finished addressing reviewer comments and indicate whether the authors have addressed all of your concerns. If you think the submission is ready to be accepted, please indicate so in a comment.

@davemlz once all reviewers are satisfied with the submission I will move on to the final steps of the JOSS process.
