### 2021-05-28 #3168 @giswqs
I would like to make two minor suggestions for the **Summary** paragraph.

- There are `e.g. Landsat, Sentinel, MODIS` and `e.g. Sentinel, Landsat and MODIS`. I would suggest making them consistent, such as as `e.g., Landsat, Sentinel, MODIS`
- `Google Earth Engine` were used three times. Since you already use the acronym `GEE` for the first one, the other two can be replaced with GEE.
