### 2021-05-28 b207484
Improved paper (https://github.com/openjournals/joss-reviews/issues/3168)
