+++ paper/paper.md
@@ -22,12 +22,12 @@ bibliography: paper.bib
 
 [Google Earth Engine](https://earthengine.google.com/) (GEE) is a cloud-based service for geospatial processing of vector and raster data [@Gorelick2017].
 The GEE platform has [JavaScript and Python APIs](https://developers.google.com/earth-engine/guides) with different methods to process geospatial objects.
-Google Earth Engine also provides a [multi-petabyte data catalog](https://developers.google.com/earth-engine/datasets/) of geospatial datasets and satellite imagery (e.g. Landsat, Sentinel, MODIS).
+GEE also provides a [multi-petabyte data catalog](https://developers.google.com/earth-engine/datasets/) of geospatial datasets and satellite imagery (e.g. Landsat, Sentinel, MODIS).
 The eemont package extends the [GEE Python API](https://developers.google.com/earth-engine/guides/python_install) with pre-processing and processing tools
-for the commonly used satellite imagery (e.g. Sentinel, Landsat and MODIS) by adding new methods for different
+for the commonly used satellite imagery (e.g. Landsat, Sentinel, MODIS) by adding new methods for different
 [Earth Engine Objects](https://developers.google.com/earth-engine/guides/objects_methods_overview) that are friendly with the Python method chaining. The package
 can be used with [geemap](https://geemap.org/) [@Wu2020] for interactive visualization in Jupyter Notebooks, with the
-[Google Earth Engine Plugin for QGIS](https://gee-community.github.io/qgis-earthengine-plugin/) for processing and visualization inside [QGIS](https://www.qgis.org/es/site/) [@QGIS_software], and it can be used in R with [rgee](https://github.com/r-spatial/rgee) [@Aybar2020].
+[GEE Plugin for QGIS](https://gee-community.github.io/qgis-earthengine-plugin/) for processing and visualization inside [QGIS](https://www.qgis.org/es/site/) [@QGIS_software], and it can be used in R with [rgee](https://github.com/r-spatial/rgee) [@Aybar2020].
 
 # Statement of need
 
@@ -106,7 +106,7 @@ L8 = (ee.ImageCollection('LANDSAT/LC08/C01/T1_SR')
     .index(['NDVI','EVI','GNDVI']))
 ```
 
-These methods support multiple datasets from the [Google Earth Engine STAC](https://earthengine-stac.storage.googleapis.com/catalog/catalog.json)
+These methods support multiple datasets from the [GEE STAC](https://earthengine-stac.storage.googleapis.com/catalog/catalog.json)
 and can be coupled with additional packages built on top of the GEE Python API such as [geemap](https://geemap.org/) [@Wu2020],
 [geetools](https://github.com/gee-community/gee_tools) [@Principe2021] and [restee](https://kmarkert.github.io/restee/) [@Markert2021], extending and simplifying the
 use of GEE.
@@ -119,7 +119,7 @@ extending the GEE Python API for the community.
 
 # Compatibility with QGIS
 
-The eemont python package can be used in QGIS with the [Google Earth Engine Plugin for QGIS](https://gee-community.github.io/qgis-earthengine-plugin/) by installing the package
+The eemont python package can be used in QGIS with the [GEE Plugin for QGIS](https://gee-community.github.io/qgis-earthengine-plugin/) by installing the package
 using the OSGeo4W Shell:
 
 ```
