### 2021-05-30 5faa4aa
Small copyedits associated with JOSS review
openjournals/joss-reviews/issues/3168
