### 2021-05-30 82ae016
Add curly braces for title capitalization
openjournals/joss-reviews/issues/3168
