### 2021-05-30 #3168 @kbarnhart
@davemlz I've made two pull requests with very small changes to the text and the .bib file. Please consider these PRs. After you have done so, please test that the article builds correctly here, and presuming it does do the following:

Know also that I'll be out of the office June 3-June 8.
