+++ setup.py
@@ -15,7 +15,7 @@ def read(filename):
 
 setup(
     name="eemont",
-    version="0.1.9",
+    version="0.2.0",
     url="https://github.com/davemlz/eemont",
     license='MIT',
 
