+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, David Montero Loaiza'
 author = 'David Montero Loaiza'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.9'
+release = '0.2.0'
 
 # -- General configuration ---------------------------------------------------
 
