### 2021-06-04 #3168 @davemlz
Hi, @kbarnhart!

The package is now in Zenodo:

- Tagged release: 0.2.0
- Software archived in: Zenodo
- List of authors revised.
- DOI: 10.5281/zenodo.4900067
