+++ docs/tutorials.rst
@@ -16,4 +16,7 @@ Here you can find a collection of eemont tutorials in Jupyter Notebooks and RMar
 11. Creating a Bounding Box From a Query (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/011-Creating-A-Bounding-Box-From-Query.ipynb>`_) 
 12. Computing Spectral Indices on MODIS (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/012-Spectral-Indices-MODIS-MOD09GA.ipynb>`_) 
 13. Time Series By Region and Conversion to Pandas (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/013-Time-Series-By-Region-Pandas.ipynb>`_) 
-14. Time Series By Regions and Conversion to Pandas (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/014-Time-Series-By-Regions-Pandas.ipynb>`_) 
+14. Time Series By Regions and Conversion to Pandas (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/014-Time-Series-By-Regions-Pandas.ipynb>`_)
+15. Scaling and Offseting ANY RASTER DATASET From the GEE STAC (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/015-Scaling-ANY-Raster-From-GEE-STAC.ipynb>`_) 
+16. Spectral Indices From the Awesome Spectral Indices for GEE (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/016-Spectral-Indices-From-Awesome-Spectral-Indices-List.ipynb>`_) 
+17. Masking Clouds and Shadows in VIIRS Products (`Jupyter Notebook <https://github.com/davemlz/eemont/blob/master/tutorials/017-VIIRS-Products-Clouds-Masking.ipynb>`_) 
