### 2021-06-07 7fa7ccd
Implement Plus Codes (#27)
* Implement plus code to lng/lat functions

* Implement ee.Geometry.Point from plus code and create placeholder methods for other geometries

* Undo autoformatting

* Undo autoformatting

* Add tests for PointFromPlusCode (#23)

* Add function for converting llnglat to plus code and split openlocationcode import into separate function (#23)

* Set default codelength and implement Geometry to plus code conversion with tests (#23)

* Implement Feature to plus code conversion with tests (#23)

* Docstrings

* Docstring formatting

* Improve the plus code splitting algor
[…]
