+++ tox.ini
@@ -3,4 +3,6 @@ envlist = py36,py37,py38,py39
 
 [testenv]
 commands = pytest tests
-deps = pytest
+deps = 
+    pytest
+    openlocationcode
