+++ docs/index.rst
@@ -82,6 +82,7 @@ Welcome to eemont!
 
 - `Overview`_
 - `Google Earth Engine Community: Developer Resources`_
+- `Additional Resources`_
 - `How does it work?`_
 - `Installation`_
 - `Features`_
@@ -98,6 +99,16 @@ Google Earth Engine Community: Developer Resources
 -----------------------------------------------------
 
 The eemont Python package can be found in the `Earth Engine Community: Developer Resources <https://developers.google.com/earth-engine/tutorials/community/developer-resources>`_ together with other awesome resources such as `geemap <https://geemap.org/>`_ and `rgee <https://github.com/r-spatial/rgee>`_.
+
+Additional Resources
+--------------------
+
+If you like eemont, you might be interested in...
+
+- `Awesome Spectral Indices for GEE <https://github.com/davemlz/awesome-ee-spectral-indices>`_: A ready-to-use curated list of spectral indices for Google Earth Engine.
+- `spectral <https://github.com/davemlz/spectral>`_: Awesome Spectral Indices for the Google Earth Engine JavaScript API (Code Editor).
+- `eeExtra <https://github.com/r-earthengine/ee_extra>`_: A ninja Python package behind rgee, rgeeExtra and eemont.
+- `rgeeExtra <https://github.com/r-earthengine/rgeeExtra>`_: High-level functions to process spatial and simple Earth Engine objects.
 
 How does it work?
 -------------------
