+++ docs/classes/stubs/eemont.geometry.MultiPointFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.MultiPointFromPlusCodes
+=======================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: MultiPointFromPlusCodes
