+++ docs/classes/stubs/eemont.geometry.MultiLineStringFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.MultiLineStringFromPlusCodes
+============================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: MultiLineStringFromPlusCodes
