+++ docs/classes/stubs/eemont.geometry.PointFromPlusCode.rst
@@ -1,0 +1,6 @@
+eemont.geometry.PointFromPlusCode
+=================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: PointFromPlusCode
