+++ docs/classes/stubs/eemont.geometry.PolygonFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.PolygonFromPlusCodes
+====================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: PolygonFromPlusCodes
