+++ docs/classes/stubs/eemont.geometry.plusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.plusCodes
+=========================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: plusCodes
