+++ docs/classes/stubs/eemont.geometry.RectangleFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.RectangleFromPlusCodes
+======================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: RectangleFromPlusCodes
