+++ docs/classes/stubs/eemont.feature.plusCodes.rst
@@ -1,0 +1,6 @@
+eemont.feature.plusCodes
+========================
+
+.. currentmodule:: eemont.feature
+
+.. autofunction:: plusCodes
