+++ docs/classes/stubs/eemont.geometry.LinearRingFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.LinearRingFromPlusCodes
+=======================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: LinearRingFromPlusCodes
