+++ docs/classes/stubs/eemont.geometry.MultiPolygonFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.MultiPolygonFromPlusCodes
+=========================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: MultiPolygonFromPlusCodes
