+++ docs/classes/stubs/eemont.geometry.LineStringFromPlusCodes.rst
@@ -1,0 +1,6 @@
+eemont.geometry.LineStringFromPlusCodes
+=======================================
+
+.. currentmodule:: eemont.geometry
+
+.. autofunction:: LineStringFromPlusCodes
